Gets the covariance of the innovations in the transition equation.

Gets the covariance of the innovations in the transition equation.

Usage

.ssf_V(x, pos = 0)

Arguments

x

A state block or a state space form.

pos

The 0-based position (use 0 for time invariant models)

Value

The requested matrix.

Examples

sarima<-.sarma(12, .9, -.6, .9, -.5)
.ssf_V(sarima)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,]  1.0000000 -1.5000000  1.3500000 -1.2150000  1.0935000 -0.9841500
#>  [2,] -1.5000000  2.2500000 -2.0250000  1.8225000 -1.6402500  1.4762250
#>  [3,]  1.3500000 -2.0250000  1.8225000 -1.6402500  1.4762250 -1.3286025
#>  [4,] -1.2150000  1.8225000 -1.6402500  1.4762250 -1.3286025  1.1957423
#>  [5,]  1.0935000 -1.6402500  1.4762250 -1.3286025  1.1957423 -1.0761680
#>  [6,] -0.9841500  1.4762250 -1.3286025  1.1957423 -1.0761680  0.9685512
#>  [7,]  0.8857350 -1.3286025  1.1957423 -1.0761680  0.9685512 -0.8716961
#>  [8,] -0.7971615  1.1957423 -1.0761680  0.9685512 -0.8716961  0.7845265
#>  [9,]  0.7174454 -1.0761680  0.9685512 -0.8716961  0.7845265 -0.7060738
#> [10,] -0.6457008  0.9685512 -0.8716961  0.7845265 -0.7060738  0.6354665
#> [11,]  0.5811307 -0.8716961  0.7845265 -0.7060738  0.6354665 -0.5719198
#> [12,] -0.5230177  0.7845265 -0.7060738  0.6354665 -0.5719198  0.5147278
#> [13,] -0.9292841  1.3939262 -1.2545335  1.1290802 -1.0161722  0.9145550
#> [14,]  1.6763557 -2.5145335  2.2630802 -2.0367722  1.8330950 -1.6497855
#>             [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,]  0.8857350 -0.7971615  0.7174454 -0.6457008  0.5811307 -0.5230177
#>  [2,] -1.3286025  1.1957423 -1.0761680  0.9685512 -0.8716961  0.7845265
#>  [3,]  1.1957423 -1.0761680  0.9685512 -0.8716961  0.7845265 -0.7060738
#>  [4,] -1.0761680  0.9685512 -0.8716961  0.7845265 -0.7060738  0.6354665
#>  [5,]  0.9685512 -0.8716961  0.7845265 -0.7060738  0.6354665 -0.5719198
#>  [6,] -0.8716961  0.7845265 -0.7060738  0.6354665 -0.5719198  0.5147278
#>  [7,]  0.7845265 -0.7060738  0.6354665 -0.5719198  0.5147278 -0.4632550
#>  [8,] -0.7060738  0.6354665 -0.5719198  0.5147278 -0.4632550  0.4169295
#>  [9,]  0.6354665 -0.5719198  0.5147278 -0.4632550  0.4169295 -0.3752366
#> [10,] -0.5719198  0.5147278 -0.4632550  0.4169295 -0.3752366  0.3377129
#> [11,]  0.5147278 -0.4632550  0.4169295 -0.3752366  0.3377129 -0.3039416
#> [12,] -0.4632550  0.4169295 -0.3752366  0.3377129 -0.3039416  0.2735475
#> [13,] -0.8230995  0.7407895 -0.6667106  0.6000395 -0.5400356  0.4860320
#> [14,]  1.4848069 -1.3363262  1.2026936 -1.0824242  0.9741818 -0.8767636
#>            [,13]      [,14]
#>  [1,] -0.9292841  1.6763557
#>  [2,]  1.3939262 -2.5145335
#>  [3,] -1.2545335  2.2630802
#>  [4,]  1.1290802 -2.0367722
#>  [5,] -1.0161722  1.8330950
#>  [6,]  0.9145550 -1.6497855
#>  [7,] -0.8230995  1.4848069
#>  [8,]  0.7407895 -1.3363262
#>  [9,] -0.6667106  1.2026936
#> [10,]  0.6000395 -1.0824242
#> [11,] -0.5400356  0.9741818
#> [12,]  0.4860320 -0.8767636
#> [13,]  0.8635689 -1.5578107
#> [14,] -1.5578107  2.8101684