Title

Title

Usage

sts_outliers(
  y,
  period = NA,
  X = NULL,
  X.td = NULL,
  level = 1,
  slope = 1,
  noise = 1,
  seasonal = c("Trigonometric", "Dummy", "Crude", "HarrisonStevens", "Fixed", "Unused"),
  ao = TRUE,
  ls = TRUE,
  so = FALSE,
  cv = 0,
  tcv = 0,
  estimation.forward = c("Score", "Point", "Full"),
  estimation.backward = c("Point", "Score", "Full")
)

Arguments

y

input time series.

period

annual frequency.

X

Regression variables (same length as y) or NULL

X.td

Groups of days for trading days regressors. The length of the array must be 7. It indicates to what group each week day belongs. The first item corresponds to Mondays and the last one to Sundays. The group used for contrasts (usually Sundays) is identified by 0. The other groups are identified by 1, 2,... n (<= 6). For instance, usual trading days are defined by c(1,2,3,4,5,6,0), week days by c(1,1,1,1,1,0,0), etc...

level

-1 = no level, 0 = fixed level, 1 = sotchastic level

seasonal

Seasonal model

ao, ls, so

boolean indicating if additive outliers (ao), level shift (ls) and seasonal outliers (so) should be detected.

Examples

 x<-rjd3toolkit::Retail$BookStores
 sts_outliers(x)
#> $model
#> $model$y
#>   [1]  790  540  536  524  553  589  593  895  863  647  642 1166  999  568  602
#>  [16]  583  613  619  608  985  905  669  693 1275 1055  636  635  610  684  726
#>  [31]  679 1156 1023  733  772 1410 1309  720  696  689  786  808  783 1248 1103
#>  [46]  747  851 1468 1375  758  743  751  855  839  787 1361 1042  899  908 1600
#>  [61] 1559  825  813  806  893  876  830 1301 1159  926  996 1771 1467  923  883
#>  [76]  854  920  939  882 1385 1247  954  991 1849 1514  990  976  934  977 1032
#>  [91] 1024 1388 1242 1020 1090 1998 1506 1091 1031  975 1105 1123 1049 1595 1407
#> [106] 1015 1100 1895 1586 1070 1063  935 1071 1078 1006 1821 1407 1016 1099 1958
#> [121] 1973 1001 1009  968 1118 1046 1038 1799 1491  998 1060 1949 2056  992  926
#> [136]  976 1109 1182 1130 2142 1574 1047 1037 2065 2148 1108 1073 1023 1112 1206
#> [151] 1172 2119 1560 1069 1088 2199 2075 1086 1101  997 1120 1154 1214 2219 1553
#> [166] 1064 1150 2256 2225 1101 1092  987 1176 1238 1135 2142 1555 1055 1143 2134
#> [181] 2222 1040 1029  931 1135 1162 1221 2330 1579 1129 1227 2179 2301 1157 1026
#> [196] 1005 1159 1077 1124 2435 1481 1032 1044 2029 2214  985  961  929 1071 1058
#> [211] 1056 2337 1548  975  979 1938 2178  962  962  871 1032 1040 1025 2183 1448
#> [226]  949 1030 1982
#> 
#> $model$variables
#> NULL
#> 
#> $model$X
#> NULL
#> 
#> $model$b
#> numeric(0)
#> 
#> $model$bcov
#> <0 x 0 matrix>
#> 
#> $model$components
#>                 [,1]      [,2]          [,3]       [,4]
#>   [1,]  -7.844919677  654.7048  6.0703502307  143.14014
#>   [2,]   2.583068956  664.4507  6.0668987792 -127.03377
#>   [3,]   0.959926251  672.9829  6.0611323235 -137.94286
#>   [4,]   0.492515849  681.0596  6.0534731927 -157.55216
#>   [5,]   0.530827222  688.8979  6.0441380742 -136.42876
#>   [6,]   1.692138707  696.4782  6.0333605107 -109.17033
#>   [7,]   2.913107048  703.2548  6.0218849759 -113.16795
#>   [8,]  -2.492001645  708.6551  6.0109931203  188.83686
#>   [9,]   0.758725454  715.2121  5.9995885629  147.02914
#>  [10,]   2.551130584  721.4022  5.9880051125  -76.95336
#>  [11,]   0.334115462  726.3855  5.9773651647  -84.71958
#>  [12,]  -3.121335287  731.2015  5.9678157171  437.91982
#>  [13,]   6.535552309  737.4705  5.9579835066  254.99399
#>  [14,]  -2.301301547  740.6675  5.9507439154 -170.36617
#>  [15,]   1.296334770  744.9355  5.9450844628 -144.23180
#>  [16,]   2.771460708  748.5904  5.9415754841 -168.36189
#>  [17,]  -0.211586693  750.9434  5.9414363116 -137.73179
#>  [18,]  -2.817456692  753.3953  5.9445738554 -131.57786
#>  [19,]  -1.635246290  757.1705  5.9497485472 -147.53522
#>  [20,]  -2.636195649  761.7169  5.9562409439  225.91926
#>  [21,]  -1.968861015  767.5050  5.9628912246  139.46382
#>  [22,]   0.035353305  774.2223  5.9688331690 -105.25762
#>  [23,]   1.204789218  780.9289  5.9740823312  -89.13366
#>  [24,]   0.592373336  787.0762  5.9791687704  487.33139
#>  [25,]  -8.221869568  792.9512  5.9843531072  270.27071
#>  [26,]   0.630228719  802.6834  5.9860180471 -167.31367
#>  [27,]   0.640036160  812.1221  5.9844408657 -177.76215
#>  [28,]  -2.359443202  821.2593  5.9799031534 -208.89989
#>  [29,]  -0.682286207  831.4975  5.9713668492 -146.81519
#>  [30,]   1.188164510  842.0468  5.9585317745 -117.23493
#>  [31,]  -1.882221842  852.0265  5.9419206744 -171.14430
#>  [32,]   3.291635579  862.8715  5.9207054462  289.83682
#>  [33,]   1.531696026  872.1531  5.8963342782  149.31518
#>  [34,]   0.328635522  880.6927  5.8694810556 -148.02133
#>  [35,]  -2.279007365  889.0514  5.8402903649 -114.77242
#>  [36,]   1.907170047  898.4488  5.8077595337  509.64408
#>  [37,]   7.644352773  906.9200  5.7727276409  394.43566
#>  [38,]   0.390767799  911.7746  5.7385578964 -192.16535
#>  [39,]  -2.259303661  916.4119  5.7054222224 -218.15261
#>  [40,]  -0.733021236  922.0747  5.6723266156 -232.34165
#>  [41,]  -0.541906585  928.0478  5.6389485759 -141.50586
#>  [42,]   2.370748600  934.2414  5.6050496860 -128.61214
#>  [43,]   3.724333298  939.2903  5.5716729806 -160.01468
#>  [44,]  -0.313161652  942.5610  5.5404570178  305.75219
#>  [45,]   6.863760305  945.9471  5.5112640188  150.18914
#>  [46,]  -7.271683248  946.0882  5.4871137643 -191.81648
#>  [47,]   1.017953387  949.6121  5.4648070039  -99.63002
#>  [48,]  -3.337370242  952.6367  5.4447915964  518.70064
#>  [49,]  -4.802769212  957.2050  5.4255992322  422.59774
#>  [50,]  -0.934386137  964.0044  5.4051168846 -205.07000
#>  [51,]  -0.890244662  971.2210  5.3829334608 -227.33079
#>  [52,]   0.527086179  978.8326  5.3586572889 -228.35970
#>  [53,]   2.434998927  986.1730  5.3325202657 -133.60796
#>  [54,]  -0.984795653  992.3463  5.3055936936 -152.36150
#>  [55,]  -2.001159189  998.9541  5.2774443016 -209.95296
#>  [56,]   9.130002831 1006.4714  5.2471916598  345.39861
#>  [57,] -10.686124270 1009.6807  5.2188526048   43.00539
#>  [58,]   6.046801495 1017.8685  5.1877256687 -124.91530
#>  [59,]  -2.255920504 1023.1920  5.1564712025 -112.93612
#>  [60,]  -3.663627508 1029.5413  5.1240966915  574.12234
#>  [61,]  15.036744441 1037.5747  5.0889902857  506.38858
#>  [62,]  -2.868016579 1038.5278  5.0577675524 -210.65979
#>  [63,]  -1.494644443 1040.7935  5.0291666789 -226.29881
#>  [64,]   0.142461843 1043.7308  5.0025300811 -237.87326
#>  [65,]  -0.099131008 1046.5748  4.9779204367 -153.47562
#>  [66,]  -1.435967956 1049.4405  4.9552941320 -172.00457
#>  [67,]  -1.458628673 1052.9565  4.9340193984 -221.49787
#>  [68,]  -7.384225116 1057.1346  4.9134544975  251.24964
#>  [69,]   1.282418599 1064.7518  4.8903506650   92.96574
#>  [70,]   0.633155989 1071.7451  4.8652721139 -146.37830
#>  [71,]   5.357320078 1078.4167  4.8384974076  -87.77403
#>  [72,]   5.607467666 1082.5514  4.8123835531  682.84109
#>  [73,]  -6.665938247 1084.0328  4.7893976122  389.63314
#>  [74,]   2.019021700 1088.6144  4.7666068383 -167.63338
#>  [75,]  -0.248521539 1092.2272  4.7448995203 -208.97863
#>  [76,]  -1.453401596 1095.9347  4.7241663186 -240.48128
#>  [77,]  -1.140655887 1100.3024  4.7037677945 -179.16178
#>  [78,]  -1.783908346 1105.1842  4.6832021051 -164.40032
#>  [79,]  -4.588094081 1110.8813  4.6616844018 -224.29317
#>  [80,]   3.510924325 1118.7064  4.6371961065  262.78264
#>  [81,]   5.671917937 1124.8622  4.6112818855  116.46593
#>  [82,]  -1.562511660 1128.3345  4.5864371565 -172.77198
#>  [83,]  -5.070627609 1132.5141  4.5619744767 -136.44345
#>  [84,]  -2.034906832 1139.0449  4.5356629723  711.98998
#>  [85,]   2.865530669 1146.5029  4.5066073656  364.63158
#>  [86,]  -0.854209511 1152.5892  4.4760683756 -161.73499
#>  [87,]   2.723438175 1159.0452  4.4436701843 -185.76864
#>  [88,]   3.007686202 1164.1928  4.4106109958 -233.20048
#>  [89,]  -3.610843447 1167.8981  4.3782140719 -187.28729
#>  [90,]   0.632536357 1173.2629  4.3448907856 -141.89540
#>  [91,]   7.711544585 1178.2979  4.3109194278 -162.00946
#>  [92,] -13.211730598 1179.6859  4.2796927701  221.52583
#>  [93,]  -9.676178941 1187.2327  4.2453981785   64.44343
#>  [94,]   2.370037304 1199.2789  4.2037785206 -181.64891
#>  [95,]   3.902441306 1210.1729  4.1558765182 -124.07539
#>  [96,]  11.502963437 1219.1907  4.1034090896  767.30632
#>  [97,]  -9.023225320 1222.7665  4.0514370795  292.25669
#>  [98,]   4.005498436 1230.5180  3.9959906291 -143.52353
#>  [99,]  -0.512417785 1236.3374  3.9388319978 -204.82497
#> [100,]   1.150846079 1242.3397  3.8797357421 -268.49051
#> [101,]   6.671671618 1247.7436  3.8192081884 -149.41531
#> [102,]   5.172953682 1249.9612  3.7601846059 -132.13417
#> [103,]   1.914723301 1249.6961  3.7049408877 -202.61081
#> [104,]  -3.372500551 1248.4786  3.6543194356  349.89390
#> [105,]   7.885578596 1248.7906  3.6068364840  150.32380
#> [106,]  -0.661428410 1245.3605  3.5659613727 -229.69910
#> [107,]   1.714033761 1242.1995  3.5314030996 -143.91349
#> [108,]  -8.225120131 1238.2008  3.5039157711  665.02436
#> [109,] -16.923027313 1238.0283  3.4798806627  364.89475
#> [110,]   4.280296288 1245.7607  3.4518523181 -180.04100
#> [111,]   7.193248738 1251.4597  3.4217138965 -195.65290
#> [112,]  -3.741624031 1253.7582  3.3926301404 -315.01659
#> [113,]  -3.024914314 1257.7808  3.3629548846 -183.75584
#> [114,]  -1.336519943 1263.1909  3.3313572875 -183.85436
#> [115,]  -4.563621498 1269.1956  3.2972493336 -258.63198
#> [116,]  12.762175951 1277.3044  3.2586232118  530.93341
#> [117,]  -5.477052248 1279.3951  3.2210937699  133.08191
#> [118,]  -0.524068251 1284.0145  3.1822513236 -267.49044
#> [119,]   0.427979021 1288.8406  3.1418653038 -190.26854
#> [120,]   2.021188478 1293.4257  3.1001240042  662.55310
#> [121,]  15.320863946 1297.0221  3.0579166671  660.65700
#> [122,]  -4.172298082 1293.3981  3.0219838609 -288.22578
#> [123,]  -0.596785268 1291.6929  2.9904899410 -282.09615
#> [124,]   3.073661265 1290.2359  2.9631723458 -325.30956
#> [125,]   2.956783289 1287.3115  2.9413833635 -172.26824
#> [126,]  -5.988145533 1282.9799  2.9264238608 -230.99173
#> [127,]  -2.389319581 1281.4390  2.9156592926 -241.04964
#> [128,] -14.896042812 1281.0068  2.9080384537  532.88929
#> [129,]   2.402283555 1287.5462  2.8970076799  201.05156
#> [130,]   0.927468552 1292.9490  2.8836238801 -295.87645
#> [131,]   1.075606131 1297.9039  2.8682951034 -238.97949
#> [132,]  -3.691215745 1302.3395  2.8514945737  650.35171
#> [133,]   4.057909330 1308.4878  2.8315983032  743.45433
#> [134,]  -4.693092082 1312.7149  2.8103916101 -316.02179
#> [135,] -11.968827688 1319.1196  2.7858097214 -381.15081
#> [136,]  -1.346180210 1331.1076  2.7525868304 -353.76139
#> [137,]  -0.382656200 1343.6930  2.7101306718 -234.31034
#> [138,]   4.510317753 1356.4153  2.6582728921 -178.92557
#> [139,]   1.076688113 1366.9724  2.5989978510 -238.04913
#> [140,]  17.176278714 1376.9659  2.5327792483  747.85783
#> [141,]   3.556399572 1378.8455  2.4671739610  191.59808
#> [142,]  -2.467538499 1378.9933  2.4037466629 -329.52574
#> [143,]  -7.387584232 1380.2337  2.3414117355 -335.84613
#> [144,]  -5.195800303 1384.8731  2.2769189368  685.32267
#> [145,]   6.556453939 1391.8824  2.2079823220  749.56112
#> [146,]   4.011784665 1395.7509  2.1374864691 -291.76268
#> [147,]   3.793437153 1397.6692  2.0671964032 -328.46267
#> [148,]   1.148142095 1397.7399  1.9987810852 -375.88809
#> [149,]  -2.379930294 1397.2043  1.9327456519 -282.82437
#> [150,]   2.847383726 1397.7177  1.8680430288 -194.56507
#> [151,]   0.312619160 1396.8323  1.8059259512 -225.14491
#> [152,]  -6.134268736 1395.7383  1.7465319589  729.39597
#> [153,]   0.334910252 1397.4590  1.6871622188  162.20608
#> [154,]   1.441748914 1398.9634  1.6279640778 -331.40518
#> [155,]  -0.001036514 1399.7331  1.5695718479 -311.73211
#> [156,]   5.603355060 1400.4450  1.5119850732  792.95168
#> [157,] -11.515923633 1398.4739  1.4576690062  688.04207
#> [158,]  -1.225861063 1401.8440  1.4015571001 -314.61811
#> [159,]   3.924969439 1405.7323  1.3431100251 -308.65731
#> [160,]  -0.685785402 1407.7233  1.2840546111 -410.03750
#> [161,]  -2.755076586 1409.9765  1.2240891397 -287.22142
#> [162,]  -8.638019850 1413.4606  1.1620014874 -250.82255
#> [163,]   5.511639488 1420.9297  1.0939912701 -212.44137
#> [164,]   7.137175261 1425.7485  1.0224833894  786.11431
#> [165,]  -3.169157490 1427.1518  0.9506179172  129.01735
#> [166,]  -3.063568014 1429.9681  0.8770005510 -362.90451
#> [167,]   0.832003000 1434.1461  0.8002834430 -284.97811
#> [168,]   7.544639370 1437.8576  0.7208326413  810.59776
#> [169,]   5.006548102 1437.9548  0.6419674861  782.03869
#> [170,]   0.846396469 1435.6273  0.5658906587 -335.47373
#> [171,]   0.773186936 1432.8273  0.4929745398 -341.60046
#> [172,]   0.441992357 1429.5920  0.4235593007 -443.03402
#> [173,]   4.512165727 1426.0803  0.3578394004 -254.59245
#> [174,]   9.889841565 1420.3887  0.2978000118 -192.27859
#> [175,]  -8.107617464 1410.0035  0.2477922702 -266.89586
#> [176,] -12.936888300 1403.3669  0.2042491478  751.57002
#> [177,]   1.511760605 1402.7480  0.1614789300  150.74022
#> [178,]  -0.405602393 1401.3781  0.1201467312 -345.97250
#> [179,]  -1.274966009 1400.1569  0.0800741025 -255.88192
#> [180,]  -4.271726484 1399.4930  0.0407001098  738.77877
#> [181,]  -1.052038498 1400.7911  0.0001453639  822.26096
#> [182,]  -5.783922537 1402.5416 -0.0420529900 -356.75764
#> [183,]  -1.904375991 1406.9598 -0.0884396471 -376.05541
#> [184,]  -5.437069670 1412.2239 -0.1398524567 -475.78681
#> [185,]  -4.126183131 1419.9840 -0.1986835124 -280.85781
#> [186,]  -0.040216232 1429.6185 -0.2667481700 -267.57829
#> [187,]   7.025544517 1439.2038 -0.3440641233 -225.22934
#> [188,]   1.042802694 1445.4201 -0.4275404129  883.53710
#> [189,]   1.007359476 1451.0643 -0.5167182473  126.92830
#> [190,]   3.976118951 1456.1474 -0.6111544287 -331.12355
#> [191,]   8.972889611 1459.2731 -0.7090996233 -241.24604
#> [192,]   3.778728840 1458.0969 -0.8066061179  717.12441
#> [193,]   3.125389542 1455.0526 -0.9020114228  842.82199
#> [194,]   9.491930277 1450.4486 -0.9939404913 -302.94057
#> [195,]  -2.094109397 1441.3055 -1.0782172529 -413.21137
#> [196,]   2.964086998 1433.0592 -1.1557630322 -431.02328
#> [197,]   3.456428761 1423.3466 -1.2252737492 -267.80303
#> [198,]  -7.873783318 1411.9451 -1.2852287134 -327.07128
#> [199,]  -1.966938044 1404.1727 -1.3390920749 -278.20572
#> [200,]  11.851946151 1397.2680 -1.3877292081 1025.88008
#> [201,]  -7.922860612 1384.7617 -1.4259257298  104.16121
#> [202,]  -1.777932107 1375.9292 -1.4571673823 -342.15130
#> [203,]  -4.349552220 1367.8986 -1.4822363848 -319.54904
#> [204,]  -2.182703627 1361.8808 -1.5030463667  669.30194
#> [205,]   0.252002360 1356.8648 -1.5205576303  856.88320
#> [206,]  -6.312907893 1351.7132 -1.5346593048 -360.40033
#> [207,]  -1.286375847 1349.5054 -1.5481288176 -387.21899
#> [208,]   1.015455099 1347.8867 -1.5615321227 -419.90219
#> [209,]  -0.947565115 1345.7789 -1.5744224603 -273.83136
#> [210,]   0.815767997 1344.1022 -1.5872167214 -286.91795
#> [211,]  -1.587234343 1342.0304 -1.5995560006 -284.44321
#> [212,]   4.787219848 1340.6900 -1.6121386176  991.52275
#> [213,]  10.072829092 1337.0941 -1.6228583868  200.83310
#> [214,]  -0.816189960 1328.7680 -1.6272836655 -352.95179
#> [215,]  -6.466252981 1320.8199 -1.6257735447 -335.35362
#> [216,]  -6.584898639 1315.9029 -1.6211729175  628.68199
#> [217,]  -0.351635942 1314.0758 -1.6163788766  864.27587
#> [218,]   0.887002122 1312.4182 -1.6115461277 -351.30517
#> [219,]   1.951028658 1310.3498 -1.6062844252 -350.30084
#> [220,]  -0.978869327 1307.3726 -1.5997353944 -435.39373
#> [221,]  -0.660140076 1304.8606 -1.5923296989 -272.20043
#> [222,]   1.778960626 1302.6652 -1.5843577737 -264.44421
#> [223,]  -0.010118504 1299.6444 -1.5750369474 -274.63428
#> [224,]  -7.313146294 1296.6376 -1.5643716716  893.67554
#> [225,]  -3.979404894 1297.0679 -1.5555794387  154.91150
#> [226,]   0.240071926 1299.3715 -1.5504110281 -350.61154
#> [227,]   4.245816460 1301.5677 -1.5487608177 -275.81353
#> [228,]   3.750818563 1301.7763 -1.5487608177  676.47286
#> 
#> $model$linearized
#>   [1]  790  540  536  524  553  589  593  895  863  647  642 1166  999  568  602
#>  [16]  583  613  619  608  985  905  669  693 1275 1055  636  635  610  684  726
#>  [31]  679 1156 1023  733  772 1410 1309  720  696  689  786  808  783 1248 1103
#>  [46]  747  851 1468 1375  758  743  751  855  839  787 1361 1042  899  908 1600
#>  [61] 1559  825  813  806  893  876  830 1301 1159  926  996 1771 1467  923  883
#>  [76]  854  920  939  882 1385 1247  954  991 1849 1514  990  976  934  977 1032
#>  [91] 1024 1388 1242 1020 1090 1998 1506 1091 1031  975 1105 1123 1049 1595 1407
#> [106] 1015 1100 1895 1586 1070 1063  935 1071 1078 1006 1821 1407 1016 1099 1958
#> [121] 1973 1001 1009  968 1118 1046 1038 1799 1491  998 1060 1949 2056  992  926
#> [136]  976 1109 1182 1130 2142 1574 1047 1037 2065 2148 1108 1073 1023 1112 1206
#> [151] 1172 2119 1560 1069 1088 2199 2075 1086 1101  997 1120 1154 1214 2219 1553
#> [166] 1064 1150 2256 2225 1101 1092  987 1176 1238 1135 2142 1555 1055 1143 2134
#> [181] 2222 1040 1029  931 1135 1162 1221 2330 1579 1129 1227 2179 2301 1157 1026
#> [196] 1005 1159 1077 1124 2435 1481 1032 1044 2029 2214  985  961  929 1071 1058
#> [211] 1056 2337 1548  975  979 1938 2178  962  962  871 1032 1040 1025 2183 1448
#> [226]  949 1030 1982
#> 
#> 
#> $bsm
#> $bsm$initial
#> $bsm$initial$level
#> [1] 0.809322
#> 
#> $bsm$initial$slope
#> [1] 0.0006285893
#> 
#> $bsm$initial$seasonal
#> [1] 0.336674
#> 
#> $bsm$initial$noise
#> [1] 1
#> 
#> $bsm$initial$tau
#>                [,1]         [,2]         [,3]         [,4]      [,5]      [,6]
#>   [1,] 4.403299e+00          NaN          NaN          NaN  0.000000  4.403299
#>   [2,] 4.565699e-01 4.403299e+00 4.403299e+00          NaN  4.403299  4.600546
#>   [3,] 5.520448e-02 1.125368e+00 2.710500e+00          NaN  4.600546  4.612089
#>   [4,] 9.446011e-03 6.098112e-01 1.959421e+00          NaN  4.612089  4.613230
#>   [5,] 1.751133e-02 4.345799e-01 1.542689e+00          NaN  4.613230  4.626308
#>   [6,] 1.939120e-01 3.023014e-01 1.270863e+00          NaN  4.626308  4.802833
#>   [7,] 5.466267e-01 7.488388e-02 9.770274e-01          NaN  4.802833  5.201797
#>   [8,] 3.802373e-01 2.394461e-02 6.462582e-01          NaN  5.201797  5.452091
#>   [9,] 3.748187e-02 3.735881e-02 5.369601e-01          NaN  5.452091  5.487847
#>  [10,] 4.075648e-01 6.989898e-03 4.367884e-01          NaN  5.487847  5.960463
#>  [11,] 3.834976e-03 6.991805e-02 3.039000e-01          NaN  5.960463  6.002423
#>  [12,] 6.308683e-01 8.654087e-02 2.095109e-01 4.403299e+00  6.002423  6.424912
#>  [13,] 1.743675e+00 1.689782e-02 1.886473e-01 3.450109e+00  6.424528  6.428568
#>  [14,] 2.020481e-01 6.011052e-01 9.557088e-02 8.604637e-02  2.479744  6.428860
#>  [15,] 5.203947e-02 2.063467e-01 5.675954e-02 8.785492e-03  2.165903  3.413090
#>  [16,] 2.906173e-01 3.589301e-01 2.449355e-02 6.075316e-04  2.286884  3.630279
#>  [17,] 1.948979e-03 8.963957e-01 1.522455e-03 4.474846e-02  2.671008  3.562632
#>  [18,] 2.946656e-01 8.369815e-01 4.671426e-03 4.271629e-02  2.670482  3.386512
#>  [19,] 9.931820e-02 3.117353e-01 1.695952e-02 8.698263e-01  2.768752  3.236130
#>  [20,] 2.472953e-01 1.239893e-01 2.775897e-02 3.106651e-01  2.419537  4.020426
#>  [21,] 1.439757e-01 7.514772e-04 2.764505e-02 9.369704e-02  3.764691  3.883878
#>  [22,] 2.800163e-05 4.847084e-02 1.947414e-02 1.573755e-01  3.785125  3.992228
#>  [23,] 6.087369e-02 4.997327e-02 1.289288e-02 3.471621e-01  3.673072  3.906777
#>  [24,] 1.411235e-02 3.918918e-03 1.106211e-02 1.469282e-01  3.865306  4.272493
#>  [25,] 2.563691e+00 2.208398e-04 1.103655e-02 3.773778e-02  3.684232  9.953358
#>  [26,] 1.576633e-02 1.057212e+00 1.807487e-11 1.168561e-01  9.058750  9.984945
#>  [27,] 1.883713e-02 8.928850e-01 8.856165e-03 3.959870e-02  8.973488 10.036566
#>  [28,] 2.110754e-01 7.300295e-01 3.140471e-02 2.738798e-01  8.778763  9.867601
#>  [29,] 1.369136e-02 1.325473e+00 8.277771e-02 6.703082e-02  7.972564  9.732654
#>  [30,] 4.771728e-02 1.502709e+00 1.628092e-01 1.766658e-01  7.909380  9.726678
#>  [31,] 1.287964e-01 1.173495e+00 2.538043e-01 1.215772e+00  7.976142  9.593139
#>  [32,] 3.891283e-01 1.728118e+00 3.902831e-01 4.454587e-01  7.808801  9.918879
#>  [33,] 7.247122e-02 8.289303e-01 4.982780e-01 4.336671e-01  6.811801  8.126512
#>  [34,] 7.540693e-03 5.417155e-01 5.925983e-01 2.509831e-02  6.779253  8.174218
#>  [35,] 1.966720e-01 4.617849e-01 6.854187e-01 5.027247e-01  7.073060  8.069733
#>  [36,] 1.355326e-01 9.320129e-01 8.319039e-01 1.539039e+00  6.980966  8.084540
#>  [37,] 2.217785e+00 5.291998e-01 9.479961e-01 3.328122e+00  6.754938  7.798306
#>  [38,] 5.606883e-03 5.430037e-02 8.963158e-01 1.245137e-01  1.534148  2.877459
#>  [39,] 1.880157e-01 7.904681e-02 8.391255e-01 1.327103e-01  1.548663  2.668649
#>  [40,] 1.566457e-02 3.064141e-06 8.305275e-01 4.002893e-02  1.108363  1.956436
#>  [41,] 1.419450e-02 6.212369e-03 8.358329e-01 3.339721e-03  1.072767  1.996504
#>  [42,] 2.087789e-01 2.416261e-02 8.543046e-01 1.555787e-02  1.121517  2.330084
#>  [43,] 5.072279e-01 1.918297e-02 8.246078e-01 3.057799e-02  1.568281  2.879740
#>  [44,] 1.057317e-02 3.556192e-01 7.245629e-01 1.106561e-01  1.931138  3.054622
#>  [45,] 1.768648e+00 2.809391e-01 6.413647e-01 2.365424e-01  2.028591  6.908701
#>  [46,] 1.922493e+00 1.916260e+00 4.555849e-01 4.344087e-01  4.867209  7.928144
#>  [47,] 4.475118e-02 2.433607e-01 3.957167e-01 3.275624e-03  6.254156  7.083449
#>  [48,] 3.733387e-01 3.956400e-01 3.266387e-01 3.318235e-02  6.173214  8.154265
#>  [49,] 1.010661e+00 5.598430e-02 3.014408e-01 2.391403e-01  7.539124 10.952044
#>  [50,] 2.564366e-02 1.669510e-01 3.405580e-01 1.618037e-02 10.179460 11.079511
#>  [51,] 1.955494e-02 2.613560e-01 3.931904e-01 5.341319e-03 10.027704 11.018837
#>  [52,] 1.019922e-02 3.609603e-01 4.600482e-01 3.778795e-02  9.834819 10.947582
#>  [53,] 2.228758e-01 2.871878e-01 5.237102e-01 3.619933e-01  9.824490 10.963928
#>  [54,] 3.893391e-02 5.434002e-02 5.513734e-01 1.352077e-02 10.003256 10.963456
#>  [55,] 1.403579e-01 1.292688e-01 5.966821e-01 8.917810e-01  9.790644 10.753749
#>  [56,] 3.041640e+00 3.594798e-01 6.777122e-01 2.840280e+00  8.960281 13.083677
#>  [57,] 4.174670e+00 2.683548e-01 6.033903e-01 4.319254e+00 11.990882 13.408273
#>  [58,] 1.350306e+00 6.257026e-01 7.126314e-01 7.219566e-01  9.785783 11.032441
#>  [59,] 2.036485e-01 2.168440e-03 7.172667e-01 5.174773e-01  9.076373  9.788037
#>  [60,] 4.971610e-01 1.126083e-01 7.651698e-01 3.073899e-01  9.322649 12.821825
#>  [61,] 8.290065e+00 6.194104e-01 8.865975e-01 1.656288e+00 11.224988 16.490118
#>  [62,] 2.967636e-01 1.116264e+00 7.210364e-01 1.167242e-02 11.702803 14.132156
#>  [63,] 8.112881e-02 5.007777e-01 6.194647e-01 2.380472e-02 12.455914 13.988173
#>  [64,] 1.206894e-03 2.758885e-01 5.489246e-01 2.432880e-02 12.699901 13.919849
#>  [65,] 1.054839e-03 2.996731e-01 4.801986e-01 9.209696e-04 12.661132 13.760035
#>  [66,] 6.274802e-02 2.772857e-01 4.185608e-01 2.615195e-03 12.611044 13.420541
#>  [67,] 6.386103e-02 1.341295e-01 3.780594e-01 1.115009e+00 12.507896 13.080844
#>  [68,] 2.075008e+00 4.182602e-02 3.561052e-01 3.573185e+00 12.438323 13.096119
#>  [69,] 6.827541e-02 5.141899e-01 4.326572e-01 8.983980e-01  9.631441 10.691939
#>  [70,] 1.054805e-02 3.023414e-01 4.961587e-01 4.535622e-01  9.262233 10.123890
#>  [71,] 1.087742e+00 2.343681e-01 5.555213e-01 9.196769e-01  9.539254 10.837333
#>  [72,] 1.166787e+00 3.348601e-02 5.316267e-01 8.006792e+00 10.317286 10.794765
#>  [73,] 1.780406e+00 7.635428e-01 4.275197e-01 3.785787e+00  6.429265  7.773243
#>  [74,] 1.546435e-01 4.178498e-04 4.246914e-01 1.886927e-02  2.419140  2.720024
#>  [75,] 2.156078e-03 7.393403e-02 3.944411e-01 2.079380e-01  1.961884  2.618026
#>  [76,] 8.746199e-02 5.866100e-02 3.684409e-01 8.558897e-03  1.999917  2.636340
#>  [77,] 3.829390e-02 2.818931e-03 3.625644e-01 5.578216e-02  2.036647  2.509862
#>  [78,] 1.162363e-01 5.187331e-03 3.696359e-01 2.833142e-03  1.912762  2.667310
#>  [79,] 8.009267e-01 8.409145e-02 3.999692e-01 4.286537e-03  2.134807  4.819674
#>  [80,] 5.689478e-01 7.432128e-01 4.979841e-01 8.042535e-02  3.556225  5.122362
#>  [81,] 1.227780e+00 1.443046e-01 5.443047e-01 6.369769e-01  4.296273  5.599039
#>  [82,] 9.531950e-02 1.078741e-01 5.033634e-01 4.589745e-02  4.777615  5.397718
#>  [83,] 1.004039e+00 1.718240e-02 4.872672e-01 8.534945e-01  4.542675  5.423942
#>  [84,] 2.071865e-01 2.594790e-01 5.493575e-01 3.025484e-01  3.915135  6.050746
#>  [85,] 3.773012e-01 6.404810e-01 6.546213e-01 9.902006e-02  4.566901  6.720609
#>  [86,] 3.282108e-02 1.662086e-01 7.115570e-01 7.714376e-02  5.984605  6.954610
#>  [87,] 2.630906e-01 2.740089e-01 7.882577e-01 4.867600e-02  6.050198  6.965054
#>  [88,] 3.325049e-01 3.827398e-02 8.176465e-01 3.495975e-02  5.956965  6.898868
#>  [89,] 4.965205e-01 2.988698e-02 7.908476e-01 1.393625e-01  6.192051  7.583345
#>  [90,] 1.205278e-02 7.696836e-02 8.329720e-01 3.321009e-01  6.750819  7.739030
#>  [91,] 2.166666e+00 4.295986e-02 8.650622e-01 2.224893e+00  6.674387  8.069216
#>  [92,] 6.184249e+00 5.378099e-01 7.526290e-01 2.540716e-03  5.668285 14.852513
#>  [93,] 3.539067e+00 7.323257e-01 8.837978e-01 2.910799e-01 11.675406 17.970065
#>  [94,] 2.034602e-01 4.234740e+00 1.242806e+00 1.938682e-01 11.673619 16.812027
#>  [95,] 5.672983e-01 3.131429e+00 1.600172e+00 1.661074e-01 12.112134 16.802674
#>  [96,] 4.846993e+00 1.659673e+00 1.888486e+00 2.432712e+00 12.683457 18.962740
#>  [97,] 2.660144e+00 1.403487e-02 1.860356e+00 1.135599e+00 15.669733 19.373825
#>  [98,] 5.289683e-01 8.531459e-01 2.079963e+00 1.425510e-01 14.950245 18.237085
#>  [99,] 2.092288e-02 2.106275e-01 2.193346e+00 7.883486e-02 15.821867 18.179042
#> [100,] 4.611863e-02 3.039929e-01 2.333670e+00 1.045931e+00 15.801215 18.927648
#> [101,] 1.653759e+00 1.715101e-01 2.441811e+00 1.565057e-01 17.033113 19.349164
#> [102,] 9.590129e-01 1.660612e-01 2.334395e+00 5.037891e-01 16.669433 20.312570
#> [103,] 1.169858e-01 1.067466e+00 2.073657e+00 5.044383e+00 17.757651 22.266508
#> [104,] 3.912341e-01 1.566749e+00 1.778608e+00 3.870619e-01 18.847674 29.491325
#> [105,] 2.414296e+00 7.260132e-01 1.590673e+00 2.763808e+00 22.808596 26.781756
#> [106,] 1.558828e-02 3.403184e+00 1.220042e+00 2.404343e-02 19.370564 25.837817
#> [107,] 9.404284e-02 3.115197e+00 9.114332e-01 9.214174e-01 18.925867 24.940456
#> [108,] 2.573450e+00 3.845148e+00 6.212930e-01 4.238340e+00 19.299238 24.021413
#> [109,] 1.028998e+01 8.760965e-01 5.023305e-01 1.347648e+00 17.060938 46.712714
#> [110,] 6.613013e-01 1.239669e+00 6.451519e-01 1.986218e-01 44.025812 47.816325
#> [111,] 1.891015e+00 3.526718e-01 7.286084e-01 3.845165e-01 45.509981 48.733598
#> [112,] 5.293858e-01 8.104577e-02 6.879194e-01 1.283561e-04 47.648880 48.754363
#> [113,] 3.641192e-01 3.245534e-02 7.134920e-01 1.124677e-02 46.838880 48.023992
#> [114,] 6.754904e-02 3.199767e-01 7.968764e-01 1.169158e-02 45.747662 47.456066
#> [115,] 7.206235e-01 5.353984e-01 9.115863e-01 9.981993e+00 45.733816 47.042927
#> [116,] 6.127740e+00 1.623196e+00 1.129736e+00 7.180595e+00 44.475101 46.384666
#> [117,] 1.156917e+00 9.451947e-02 1.074988e+00 6.023510e-03 34.805448 37.336931
#> [118,] 1.309918e-02 1.441987e-01 1.142897e+00 6.367887e-02 35.826070 37.306532
#> [119,] 3.788069e-03 2.050792e-01 1.226612e+00 7.126529e-01 35.618517 37.369371
#> [120,] 1.834768e-01 1.710129e-01 1.305673e+00 1.276436e+01 36.401328 37.397561
#> [121,] 8.591076e+00 1.956598e-02 1.333060e+00 2.457397e+01 36.738012 38.609942
#> [122,] 6.206104e-01 3.002357e+00 1.015216e+00 7.202798e-03 10.193723 15.443021
#> [123,] 1.241639e-02 1.511501e+00 8.159294e-01 1.863627e+00  7.753706 12.839672
#> [124,] 3.731892e-01 1.341539e+00 6.480793e-01 1.121459e-02  7.101752  9.846697
#> [125,] 3.054407e-01 2.398026e+00 4.537909e-01 1.852248e-01  7.419176 11.104862
#> [126,] 1.266546e+00 3.616353e+00 2.624047e-01 2.006646e-01  7.572357 11.253743
#> [127,] 2.011411e-01 1.398667e+00 1.696554e-01 4.348185e-04  8.402403 10.257065
#> [128,] 8.549863e+00 8.030281e-01 1.127921e-01 4.528643e-01  8.805924 26.242615
#> [129,] 2.022777e-01 9.449743e-01 1.750849e-01 9.703362e-01 25.169375 26.417757
#> [130,] 4.395301e-02 4.688799e-01 2.271818e-01 1.751521e-01 24.977334 26.008380
#> [131,] 5.462247e-02 3.033844e-01 2.740336e-01 2.478922e-01 24.585098 26.807731
#> [132,] 4.429489e-01 1.611857e-01 3.109860e-01 1.463740e+00 24.987268 26.739601
#> [133,] 4.501992e-01 6.835033e-01 3.943666e-01 2.091383e+00 25.252253 27.016105
#> [134,] 7.696174e-01 1.584249e-01 4.381252e-01 5.233616e-01 24.463036 27.091807
#> [135,] 5.109517e+00 9.189146e-01 5.527833e-01 4.667431e-01 24.321966 30.293281
#> [136,] 7.673573e-02 5.774515e+00 8.984971e-01 2.593568e-03 20.066342 27.613318
#> [137,] 4.076458e-03 6.656843e+00 1.362890e+00 1.331570e+00 19.630041 27.623611
#> [138,] 7.254204e-01 6.869346e+00 1.933611e+00 3.653564e-02 19.655439 27.533824
#> [139,] 4.751092e-02 4.312793e+00 2.457016e+00 5.560446e+00 19.634165 25.235815
#> [140,] 1.096909e+01 3.753846e+00 3.002211e+00 6.311686e+00 19.610968 24.675118
#> [141,] 4.370675e-01 3.205820e-02 2.951202e+00 5.751236e-01  5.800131  8.950998
#> [142,] 2.415582e-01 3.618787e-01 2.779506e+00 1.946856e-01  5.855781  8.938465
#> [143,] 1.951275e+00 8.263790e-02 2.700124e+00 1.814140e+00  5.497203  8.682145
#> [144,] 1.002312e+00 3.664046e-01 2.873931e+00 2.702345e-03  4.193328 11.312995
#> [145,] 1.589209e+00 1.550591e+00 3.246855e+00 2.426501e-04  6.800879 11.895043
#> [146,] 6.033473e-01 1.931614e-01 3.385568e+00 3.392188e-01  8.314811 11.668837
#> [147,] 5.591182e-01 3.247295e-03 3.370261e+00 1.414679e+00  8.286295 11.679605
#> [148,] 4.916275e-02 2.861888e-01 3.207487e+00 6.359550e-02  6.662540 10.872775
#> [149,] 1.892482e-01 4.579638e-01 3.006689e+00 8.382797e-02  6.778752 10.443722
#> [150,] 3.147618e-01 1.589257e-01 2.892524e+00 1.658057e-01  6.365549 10.360034
#> [151,] 1.269475e-03 5.735879e-01 2.679436e+00 3.170297e-01  7.297198 10.192192
#> [152,] 1.577368e+00 6.086740e-01 2.468332e+00 9.164962e-05  7.227069 11.663826
#> [153,] 4.834388e-03 5.202867e-04 2.477114e+00 1.087850e-01  8.432333 11.316162
#> [154,] 9.521865e-02 4.686351e-04 2.474109e+00 1.710431e-01  8.543780 10.956797
#> [155,] 1.865862e-03 4.794897e-02 2.418486e+00 9.472859e-01  8.728347 11.158796
#> [156,] 1.110197e+00 6.081696e-02 2.356334e+00 4.784712e+00  8.624121 11.348697
#> [157,] 4.891922e+00 8.473170e-01 2.123559e+00 2.231843e-01  5.420045 13.727879
#> [158,] 5.873851e-02 2.441154e-01 2.251896e+00 3.718758e-01  7.762918 11.224310
#> [159,] 5.371050e-01 4.213891e-01 2.425626e+00 4.799425e-01  7.132682 10.561609
#> [160,] 1.594637e-02 3.253583e-02 2.478039e+00 1.353322e-01  7.346012 10.539773
#> [161,] 2.729279e-01 6.821543e-02 2.553466e+00 4.236403e-01  6.983513 10.830655
#> [162,] 2.849888e+00 3.545753e-01 2.724440e+00 8.158454e-01  7.241350 13.547730
#> [163,] 1.149840e+00 2.808117e+00 3.226816e+00 4.307842e-01  6.569787 12.913257
#> [164,] 2.036021e+00 9.795034e-01 3.547133e+00 1.590057e+00  8.520230 12.954831
#> [165,] 3.748206e-01 5.863074e-03 3.579732e+00 1.084457e-02  8.179001 11.866537
#> [166,] 3.501973e-01 2.194558e-01 3.742690e+00 4.410616e-01  7.889038 11.880572
#> [167,] 2.425277e-02 7.180127e-01 4.040668e+00 3.718809e-04  7.623162 11.894119
#> [168,] 1.970047e+00 5.591419e-01 4.315551e+00 8.020421e-01  7.391471 13.044297
#> [169,] 9.865467e-01 2.277502e-02 4.273160e+00 1.923605e+00  9.324871 13.058221
#> [170,] 3.214223e-02 6.194469e-01 4.007711e+00 9.667463e-02  6.133551 11.179770
#> [171,] 1.658855e-02 8.136175e-01 3.712169e+00 1.401091e-02  6.272347 11.374899
#> [172,] 1.085406e-02 9.695687e-01 3.401204e+00 5.503945e-01  6.472614 11.087355
#> [173,] 7.799565e-01 1.105846e+00 3.083045e+00 2.022492e-01  6.791452 11.187552
#> [174,] 3.633902e+00 2.617173e+00 2.617883e+00 2.900132e+00  6.185372 14.405254
#> [175,] 2.500976e+00 8.065555e+00 1.885158e+00 1.034066e-01  4.873886 14.405774
#> [176,] 6.084291e+00 3.336741e+00 1.478557e+00 4.411639e-01  6.260981 16.966485
#> [177,] 9.440310e-02 6.017746e-02 1.433094e+00 7.195906e-03 12.726443 15.042293
#> [178,] 3.675056e-03 1.957547e-01 1.347993e+00 1.731690e-03 12.642850 15.410662
#> [179,] 7.579263e-02 1.629384e-01 1.273193e+00 1.342905e+00 13.069496 16.121879
#> [180,] 6.972983e-01 5.156899e-02 1.234959e+00 1.646018e+00 13.733253 16.150309
#> [181,] 5.024816e-02 9.539980e-02 1.303591e+00 3.824884e-01 11.310057 14.579636
#> [182,] 1.236165e+00 2.051054e-01 1.405068e+00 6.930611e-02 11.891866 17.313093
#> [183,] 1.196294e-01 1.362167e+00 1.671879e+00 2.005227e-02 12.386121 17.049957
#> [184,] 1.039577e+00 1.931276e+00 2.020560e+00 4.155208e-05 11.119606 18.864715
#> [185,] 6.642721e-01 4.183896e+00 2.591481e+00 2.664973e-01 10.430618 20.288104
#> [186,] 1.383108e-05 6.605929e+00 3.411458e+00 2.527633e+00  9.280359 20.365211
#> [187,] 1.860057e+00 6.623378e+00 4.354563e+00 1.339322e+00  9.981974 20.080912
#> [188,] 4.016006e-02 2.880223e+00 5.060645e+00 3.820759e+00 11.717483 25.590419
#> [189,] 4.318735e-02 2.461391e+00 5.773371e+00 1.867488e-01  9.620511 18.396745
#> [190,] 5.683356e-01 2.061501e+00 6.481895e+00 2.594413e-02 10.146891 18.393685
#> [191,] 2.918041e+00 9.039131e-01 7.005723e+00 3.973771e-01 10.587953 20.146869
#> [192,] 5.326154e-01 2.251643e-02 7.012424e+00 4.987569e-01 14.016338 20.722465
#> [193,] 3.421471e-01 3.855018e-01 6.794637e+00 4.822671e-02 15.489189 20.757577
#> [194,] 3.350170e+00 9.973753e-01 6.402932e+00 9.730069e-01 14.683699 23.065469
#> [195,] 1.444210e-01 4.756812e+00 5.489933e+00 6.047809e-01 13.702687 22.440376
#> [196,] 3.221271e-01 3.750722e+00 4.740356e+00 3.485324e-02 13.801054 21.732517
#> [197,] 4.646956e-01 5.311543e+00 3.900193e+00 1.644044e+00 12.293259 22.253877
#> [198,] 2.234815e+00 7.547095e+00 2.985551e+00 8.923707e-02 11.545989 22.348085
#> [199,] 1.435219e-01 3.174106e+00 2.468685e+00 5.088220e+00 12.127349 19.093134
#> [200,] 4.875624e+00 2.364931e+00 2.063757e+00 3.848180e+00 11.727228 17.796647
#> [201,] 2.323919e+00 8.831383e+00 1.343031e+00 6.955548e-03  5.030406 17.320958
#> [202,] 1.207594e-01 3.942546e+00 9.499241e-01 5.475043e-01  9.474014 15.404885
#> [203,] 6.625732e-01 3.106205e+00 6.553766e-01 8.605915e-02  9.243330 13.513641
#> [204,] 1.534265e-01 1.525112e+00 4.830327e-01 7.106566e-01  7.109263  9.644784
#> [205,] 2.132486e-03 9.629057e-01 3.651825e-01 9.143227e-01  5.929505  6.991058
#> [206,] 1.521879e+00 1.023687e+00 2.579830e-01 1.210935e+00  5.926713  6.975518
#> [207,] 5.515125e-02 4.300659e-02 2.448280e-01 1.367687e-02  4.276267  4.810704
#> [208,] 3.498248e-02 2.939607e-03 2.483520e-01 8.668964e-02  4.509343  4.854377
#> [209,] 3.624674e-02 3.108050e-02 2.390096e-01 5.971062e-01  4.586370  4.845791
#> [210,] 3.374438e-02 2.715385e-03 2.438453e-01 5.699914e-01  4.488863  4.853231
#> [211,] 8.807450e-02 2.964770e-02 2.359397e-01 3.478392e-02  4.293944  4.663629
#> [212,] 7.788392e-01 4.665505e-04 2.516637e-01 2.439946e+00  4.221515  8.016691
#> [213,] 3.833866e+00 3.109017e-01 2.016391e-01 1.426829e+00  7.250245 10.040245
#> [214,] 2.135992e-02 3.440475e+00 5.373341e-02 8.974521e-02  4.764917  8.706385
#> [215,] 1.544828e+00 3.170046e+00 9.957807e-05 1.108734e-02  4.765057  8.498889
#> [216,] 1.601076e+00 9.734320e-01 1.508582e-02 1.849290e-01  6.261821  7.826571
#> [217,] 1.265006e-02 1.721646e-02 2.346024e-02 6.312735e-02  6.890447  6.961312
#> [218,] 6.382468e-02 5.278091e-03 3.257914e-02 2.515768e-02  6.935943  6.935943
#> [219,] 2.334234e-01 4.446428e-02 5.641221e-02 2.658603e-01  6.924738  6.924738
#> [220,] 7.082634e-02 2.306418e-01 1.297141e-01 3.796231e-03  6.808666  6.808666
#> [221,] 3.124924e-02 1.147601e-01 2.387504e-01 1.770599e-01  6.626292  6.626292
#> [222,] 1.892445e-01 5.995392e-02 4.075798e-01 9.181379e-02  6.555731  6.555731
#> [223,] 2.718575e-05 2.549606e-01 8.673316e-01 1.693540e+00  6.505574  6.505574
#> [224,] 3.617166e+00 2.708382e-01 1.918493e+00 3.103716e-01  6.488152  6.488152
#> [225,] 1.208886e+00 4.322010e-01 2.556513e+00 6.177010e-01  3.350346  3.350346
#> [226,] 2.676006e-03 2.133994e+00 2.187141e+00 5.283584e-01  2.741148  2.741148
#> [227,] 1.287194e+00 2.586479e+00 1.017522e+00 7.901865e-03  2.592679  2.592679
#> [228,] 1.017522e+00 1.017522e+00          NaN 1.017522e+00  1.017522  1.017522
#> 
#> 
#> $bsm$final
#> $bsm$final$level
#> [1] 0.4685295
#> 
#> $bsm$final$slope
#> [1] 0.0004399601
#> 
#> $bsm$final$seasonal
#> [1] 0.185928
#> 
#> $bsm$final$noise
#> [1] 1
#> 
#> $bsm$final$tau
#>                [,1]         [,2]         [,3]         [,4]      [,5]      [,6]
#>   [1,] 4.437989e+00          NaN          NaN          NaN  0.000000  4.437989
#>   [2,] 4.746476e-01 4.437989e+00 4.437989e+00          NaN  4.437989  4.637854
#>   [3,] 6.330289e-02 1.121927e+00 2.725333e+00          NaN  4.637854  4.645872
#>   [4,] 1.658844e-02 5.904623e-01 1.954398e+00          NaN  4.645872  4.649543
#>   [5,] 1.898129e-02 3.949708e-01 1.515451e+00          NaN  4.649543  4.660143
#>   [6,] 1.926021e-01 2.675794e-01 1.233059e+00          NaN  4.660143  4.837510
#>   [7,] 5.673938e-01 5.842165e-02 9.369324e-01          NaN  4.837510  5.263855
#>   [8,] 4.150548e-01 3.906365e-02 6.061777e-01          NaN  5.263855  5.536611
#>   [9,] 3.840398e-02 2.900098e-02 5.004422e-01          NaN  5.536611  5.573762
#>  [10,] 4.341653e-01 3.431416e-03 4.034632e-01          NaN  5.573762  6.081190
#>  [11,] 7.445666e-03 9.294202e-02 2.734318e-01          NaN  6.081190  6.136391
#>  [12,] 6.498030e-01 1.212592e-01 1.808062e-01 4.437989e+00  6.136391  6.561529
#>  [13,] 1.598805e+00 7.960641e-03 1.600049e-01 3.488604e+00  6.561287  6.572989
#>  [14,] 1.969883e-01 6.064469e-01 7.655060e-02 8.432284e-02  2.677765  6.573026
#>  [15,] 6.156382e-02 2.118204e-01 4.266027e-02 6.947732e-03  2.331761  3.577144
#>  [16,] 2.809479e-01 3.822776e-01 1.523514e-02 4.660952e-05  2.472155  3.832862
#>  [17,] 1.628302e-03 9.221735e-01 2.255489e-05 4.268425e-02  2.871522  3.769427
#>  [18,] 2.886250e-01 8.659159e-01 1.087475e-02 4.714738e-02  2.870076  3.592753
#>  [19,] 9.704078e-02 3.329965e-01 2.823566e-02 9.224943e-01  2.956111  3.451878
#>  [20,] 2.521904e-01 1.390740e-01 4.259729e-02 3.335811e-01  2.608722  4.227307
#>  [21,] 1.406044e-01 1.993593e-03 4.298126e-02 1.013529e-01  3.968895  4.080878
#>  [22,] 4.533440e-05 4.010320e-02 3.307968e-02 1.550567e-01  3.983568  4.218321
#>  [23,] 5.264811e-02 3.833658e-02 2.494506e-02 3.766124e-01  3.892832  4.142687
#>  [24,] 1.272710e-02 2.113944e-03 2.267766e-02 1.032997e-01  4.079051  4.490116
#>  [25,] 2.444051e+00 7.645626e-04 2.285548e-02 7.939105e-02  3.896100 10.182818
#>  [26,] 1.436036e-02 9.840713e-01 2.293868e-03 1.222496e-01  9.362624 10.204669
#>  [27,] 1.480988e-02 8.317780e-01 2.008613e-03 3.184266e-02  9.270236 10.258059
#>  [28,] 2.012571e-01 6.912253e-01 1.626239e-02 2.719043e-01  9.054152 10.075189
#>  [29,] 1.682774e-02 1.257147e+00 5.640475e-02 6.543125e-02  8.258617  9.943838
#>  [30,] 5.103107e-02 1.449552e+00 1.251933e-01 1.810852e-01  8.191327  9.930001
#>  [31,] 1.280545e-01 1.116156e+00 2.061764e-01 1.286719e+00  8.268346  9.783524
#>  [32,] 3.916252e-01 1.656741e+00 3.310884e-01 4.813307e-01  8.091053 10.102485
#>  [33,] 8.479698e-02 7.773280e-01 4.306210e-01 4.519489e-01  7.046977  8.237434
#>  [34,] 3.903577e-03 4.802448e-01 5.157673e-01 3.440875e-02  7.027026  8.282192
#>  [35,] 1.877256e-01 4.254616e-01 6.018056e-01 5.188320e-01  7.269737  8.154137
#>  [36,] 1.314652e-01 8.678871e-01 7.386168e-01 1.537318e+00  7.181453  8.174379
#>  [37,] 2.112062e+00 4.861276e-01 8.471341e-01 3.419372e+00  6.917935  7.885995
#>  [38,] 5.519028e-03 5.769796e-02 7.977052e-01 1.234501e-01  1.685309  2.949773
#>  [39,] 1.844903e-01 8.290932e-02 7.430208e-01 1.243804e-01  1.692501  2.728502
#>  [40,] 1.942032e-02 1.243399e-04 7.346721e-01 5.037630e-02  1.230656  1.993901
#>  [41,] 1.061379e-02 6.171957e-03 7.411200e-01 5.686899e-03  1.181853  2.003320
#>  [42,] 2.031379e-01 2.097068e-02 7.585761e-01 1.395719e-02  1.222338  2.340294
#>  [43,] 5.013201e-01 2.106011e-02 7.301357e-01 4.760350e-02  1.671787  2.924803
#>  [44,] 3.544492e-03 3.603122e-01 6.344036e-01 1.460253e-01  2.074649  3.061054
#>  [45,] 1.702706e+00 3.155992e-01 5.513860e-01 2.108959e-01  2.106092  6.881950
#>  [46,] 1.911105e+00 1.959777e+00 3.751482e-01 4.188933e-01  4.885827  7.836310
#>  [47,] 3.745154e-02 2.617513e-01 3.183139e-01 4.388585e-04  6.185054  6.947807
#>  [48,] 4.025523e-01 4.041468e-01 2.549678e-01 5.017878e-02  6.108083  8.074954
#>  [49,] 8.336750e-01 5.211508e-02 2.333071e-01 3.189784e-01  7.551508 10.677185
#>  [50,] 3.155483e-02 1.279549e-01 2.645319e-01 1.389712e-02 10.016857 10.803719
#>  [51,] 2.864385e-02 2.223913e-01 3.089925e-01 6.793293e-03  9.866616 10.734699
#>  [52,] 1.004097e-02 3.364324e-01 3.685869e-01 3.633081e-02  9.659444 10.639666
#>  [53,] 2.142937e-01 2.658842e-01 4.256861e-01 3.494546e-01  9.647804 10.658971
#>  [54,] 3.505126e-02 4.784589e-02 4.502343e-01 1.477071e-02  9.827583 10.658876
#>  [55,] 1.447352e-01 1.147202e-01 4.904655e-01 8.706563e-01  9.617262 10.450590
#>  [56,] 3.012675e+00 3.392633e-01 5.647799e-01 2.732175e+00  8.793320 12.809611
#>  [57,] 4.127152e+00 2.807382e-01 4.941832e-01 4.176660e+00 11.770094 13.108598
#>  [58,] 1.321482e+00 5.956809e-01 5.946088e-01 6.733881e-01  9.573767 10.748250
#>  [59,] 1.839319e-01 1.246115e-03 5.979927e-01 5.628533e-01  8.904730  9.479255
#>  [60,] 4.851006e-01 9.609070e-02 6.401156e-01 2.337530e-01  9.121875 12.535244
#>  [61,] 8.171778e+00 5.715081e-01 7.510483e-01 1.561863e+00 11.061230 16.289015
#>  [62,] 2.972843e-01 1.154703e+00 5.928421e-01 1.211541e-02 11.580646 13.904141
#>  [63,] 8.073917e-02 5.261392e-01 4.964931e-01 2.840614e-02 12.330090 13.748935
#>  [64,] 7.335089e-04 2.952498e-01 4.298545e-01 2.167436e-02 12.581744 13.676291
#>  [65,] 3.551625e-04 3.143263e-01 3.662964e-01 2.069628e-04 12.542403 13.527711
#>  [66,] 7.452421e-02 3.008856e-01 3.091383e-01 3.414034e-03 12.497075 13.194689
#>  [67,] 7.689484e-02 1.397025e-01 2.728979e-01 1.077180e+00 12.387039 12.844850
#>  [68,] 1.970686e+00 3.852676e-02 2.546312e-01 3.444197e+00 12.308877 12.829005
#>  [69,] 5.943838e-02 4.928120e-01 3.209596e-01 8.448597e-01  9.537438 10.462289
#>  [70,] 1.448867e-02 2.980731e-01 3.776995e-01 4.218977e-01  9.204615  9.953904
#>  [71,] 1.037295e+00 2.198775e-01 4.300147e-01 9.989884e-01  9.500285 10.681723
#>  [72,] 1.136424e+00 3.337326e-02 4.086008e-01 7.996114e+00 10.254421 10.630878
#>  [73,] 1.605940e+00 7.475574e-01 3.162528e-01 3.688700e+00  6.339845  7.573532
#>  [74,] 1.473291e-01 2.910016e-03 3.106046e-01 1.771800e-02  2.510596  2.719345
#>  [75,] 2.232208e-03 8.967190e-02 2.815190e-01 2.101478e-01  2.057097  2.599817
#>  [76,] 7.634450e-02 7.247855e-02 2.566008e-01 1.090855e-02  2.095575  2.601783
#>  [77,] 4.702358e-02 8.554566e-03 2.481848e-01 6.304921e-02  2.122352  2.473852
#>  [78,] 1.150142e-01 2.134011e-03 2.520799e-01 3.899029e-03  1.989018  2.599676
#>  [79,] 7.608006e-01 6.920735e-02 2.757639e-01 1.289431e-04  2.211467  4.627488
#>  [80,] 4.455014e-01 6.737765e-01 3.569239e-01 1.032192e-01  3.575853  4.866083
#>  [81,] 1.162695e+00 1.552351e-01 3.994528e-01 6.467368e-01  4.195346  5.315219
#>  [82,] 8.823736e-02 8.732118e-02 3.669484e-01 3.495309e-02  4.619980  5.094701
#>  [83,] 9.292431e-01 1.114226e-02 3.555557e-01 9.124092e-01  4.378206  5.089360
#>  [84,] 1.496563e-01 2.609147e-01 4.111202e-01 4.284017e-01  3.742243  5.639372
#>  [85,] 2.967672e-01 5.747536e-01 5.011056e-01 1.465192e-01  4.306826  6.211202
#>  [86,] 2.637148e-02 1.679436e-01 5.533293e-01 7.203764e-02  5.592846  6.398590
#>  [87,] 2.680654e-01 2.638072e-01 6.224912e-01 6.006138e-02  5.648492  6.405995
#>  [88,] 3.269420e-01 3.334354e-02 6.478958e-01 2.554274e-02  5.543699  6.357136
#>  [89,] 4.712191e-01 3.346998e-02 6.219686e-01 1.332212e-01  5.810418  6.986864
#>  [90,] 1.446027e-02 6.548401e-02 6.578200e-01 3.124488e-01  6.337478  7.103592
#>  [91,] 2.149257e+00 3.204807e-02 6.834356e-01 2.024116e+00  6.246265  7.498012
#>  [92,] 6.308483e+00 5.748167e-01 5.772880e-01 3.765463e-03  5.266176 14.603489
#>  [93,] 3.383867e+00 7.181518e-01 6.960994e-01 2.273190e-01 11.674320 17.668173
#>  [94,] 2.030092e-01 4.093861e+00 1.024953e+00 1.885421e-01 11.751271 16.555033
#>  [95,] 5.503997e-01 3.011204e+00 1.357403e+00 1.744524e-01 12.196035 16.550059
#>  [96,] 4.782168e+00 1.590183e+00 1.628109e+00 2.317185e+00 12.793674 18.741202
#>  [97,] 2.942584e+00 1.872481e-02 1.597174e+00 1.084559e+00 15.679625 19.474902
#>  [98,] 5.798537e-01 9.209415e-01 1.817510e+00 1.332873e-01 15.206823 18.353765
#>  [99,] 9.489726e-03 2.236412e-01 1.931148e+00 8.726423e-02 16.150387 18.251026
#> [100,] 4.786739e-02 2.864064e-01 2.063957e+00 1.024352e+00 16.138547 18.999295
#> [101,] 1.608698e+00 1.562771e-01 2.164820e+00 1.492802e-01 17.352895 19.452153
#> [102,] 9.671246e-01 1.725467e-01 2.058292e+00 4.823137e-01 17.017613 20.475740
#> [103,] 1.325004e-01 1.089866e+00 1.802882e+00 5.208021e+00 18.125803 22.479025
#> [104,] 4.110638e-01 1.629762e+00 1.513637e+00 4.725094e-01 19.231561 29.688022
#> [105,] 2.247358e+00 7.513687e-01 1.331631e+00 2.743026e+00 23.343414 27.037312
#> [106,] 1.581142e-02 3.330602e+00 9.867051e-01 2.652862e-02 20.055738 26.196367
#> [107,] 1.061803e-01 3.043678e+00 7.052436e-01 8.614203e-01 19.581789 25.357439
#> [108,] 2.445061e+00 3.813732e+00 4.461391e-01 4.461736e+00 19.989678 24.439894
#> [109,] 1.035050e+01 9.090523e-01 3.410924e-01 1.601582e+00 17.766198 47.591535
#> [110,] 6.621446e-01 1.216296e+00 4.638246e-01 1.891283e-01 45.193664 48.757977
#> [111,] 1.870059e+00 3.396122e-01 5.362696e-01 3.479991e-01 46.724593 49.793904
#> [112,] 5.059708e-01 8.484270e-02 4.993795e-01 1.193447e-03 48.956729 49.798691
#> [113,] 3.306978e-01 2.668658e-02 5.198887e-01 1.463965e-02 48.117274 49.027744
#> [114,] 6.455889e-02 2.818668e-01 5.894208e-01 1.382101e-02 47.112045 48.548265
#> [115,] 7.527043e-01 4.806773e-01 6.867974e-01 1.035747e+01 47.106689 48.155038
#> [116,] 5.886465e+00 1.557069e+00 8.808129e-01 7.186840e+00 45.851779 47.478619
#> [117,] 1.084174e+00 9.173680e-02 8.315224e-01 2.039610e-02 35.705130 38.022461
#> [118,] 9.926150e-03 1.314975e-01 8.907490e-01 6.642993e-02 36.779612 38.037023
#> [119,] 6.619880e-03 1.817362e-01 9.629883e-01 6.918407e-01 36.559469 38.152044
#> [120,] 1.476450e-01 1.401029e-01 1.028755e+00 1.333999e+01 37.381422 38.164088
#> [121,] 8.483434e+00 1.656661e-02 1.051916e+00 2.519987e+01 37.709279 39.348928
#> [122,] 6.291524e-01 3.003017e+00 7.624605e-01 6.544663e-03 11.027419 15.980872
#> [123,] 1.287186e-02 1.502959e+00 5.857643e-01 1.812501e+00  8.493723 13.265350
#> [124,] 3.414422e-01 1.330430e+00 4.407512e-01 8.530258e-03  7.799085 10.371659
#> [125,] 3.159688e-01 2.331530e+00 2.804313e-01 1.763609e-01  8.157681 11.738677
#> [126,] 1.295955e+00 3.557866e+00 1.322014e-01 2.059804e-01  8.388514 11.862968
#> [127,] 2.063259e-01 1.342366e+00 6.846192e-02 2.264241e-02  9.149406 10.833398
#> [128,] 8.019496e+00 7.539072e-01 3.431793e-02 3.011139e-01  9.537798 26.557742
#> [129,] 2.085709e-01 8.870076e-01 7.191057e-02 1.040044e+00 25.785045 26.767549
#> [130,] 3.108878e-02 4.223738e-01 1.058792e-01 1.817456e-01 25.578461 26.346094
#> [131,] 4.181305e-02 2.885901e-01 1.389135e-01 3.021313e-01 25.157404 27.139125
#> [132,] 4.924297e-01 1.652465e-01 1.669011e-01 1.683404e+00 25.494504 27.134990
#> [133,] 5.951275e-01 7.311426e-01 2.341247e-01 2.459720e+00 25.825904 27.376304
#> [134,] 7.960192e-01 1.310108e-01 2.660403e-01 5.098027e-01 24.768436 27.053723
#> [135,] 5.177363e+00 8.691499e-01 3.575504e-01 4.849399e-01 24.561525 30.187177
#> [136,] 6.549556e-02 5.696895e+00 6.532732e-01 5.135751e-03 20.167058 27.354511
#> [137,] 5.292038e-03 6.504818e+00 1.067144e+00 1.354661e+00 19.761275 27.368818
#> [138,] 7.352242e-01 6.744427e+00 1.592574e+00 3.362930e-02 19.777193 27.276670
#> [139,] 4.189724e-02 4.198007e+00 2.081398e+00 5.593239e+00 19.775129 24.964357
#> [140,] 1.066262e+01 3.679070e+00 2.598489e+00 6.321330e+00 19.744624 24.381924
#> [141,] 4.571163e-01 2.870516e-02 2.551517e+00 5.756016e-01  5.730011  8.526150
#> [142,] 2.200562e-01 3.620127e-01 2.385855e+00 2.047984e-01  5.799447  8.510565
#> [143,] 1.972471e+00 9.106971e-02 2.305352e+00 1.942433e+00  5.441342  8.237821
#> [144,] 9.756874e-01 3.553921e-01 2.468833e+00 5.573290e-03  4.054799 10.772412
#> [145,] 1.553617e+00 1.507278e+00 2.822132e+00 3.547686e-04  6.692589 11.279039
#> [146,] 5.816760e-01 1.855804e-01 2.952750e+00 3.298073e-01  8.045211 11.019266
#> [147,] 5.200819e-01 3.232741e-03 2.937138e+00 1.403147e+00  8.042718 11.023451
#> [148,] 4.764281e-02 2.683183e-01 2.784171e+00 5.869598e-02  6.564043 10.336530
#> [149,] 2.047079e-01 4.324222e-01 2.595450e+00 8.851478e-02  6.691567  9.894584
#> [150,] 2.930206e-01 1.356334e-01 2.493386e+00 1.802850e-01  6.242581  9.749236
#> [151,] 3.532137e-03 5.104698e-01 2.299713e+00 2.359089e-01  7.143621  9.594059
#> [152,] 1.359978e+00 5.662737e-01 2.104083e+00 1.052541e-02  7.050781 11.015070
#> [153,] 4.053810e-03 4.492015e-05 2.104047e+00 9.483751e-02  8.290183 10.797096
#> [154,] 7.512524e-02 2.249149e-03 2.093686e+00 1.752524e-01  8.393108 10.448874
#> [155,] 3.882908e-08 4.961445e-02 2.038919e+00 1.011270e+00  8.574097 10.646825
#> [156,] 1.134758e+00 4.956391e-02 1.984976e+00 4.646576e+00  8.522090 10.809574
#> [157,] 4.792968e+00 8.173696e-01 1.767722e+00 1.864010e-01  5.286804 13.190948
#> [158,] 5.431122e-02 2.464480e-01 1.888617e+00 3.757086e-01  7.704410 10.736062
#> [159,] 5.567751e-01 4.167601e-01 2.051479e+00 5.128712e-01  7.091130 10.105599
#> [160,] 1.699745e-02 2.828810e-02 2.097016e+00 1.462549e-01  7.304056 10.077859
#> [161,] 2.743310e-01 6.331647e-02 2.165014e+00 3.944730e-01  6.932571 10.307385
#> [162,] 2.696721e+00 3.443621e-01 2.324247e+00 7.628680e-01  7.172572 12.850693
#> [163,] 1.097917e+00 2.682539e+00 2.793019e+00 4.301091e-01  6.459107 12.209033
#> [164,] 1.841030e+00 9.357574e-01 3.092643e+00 1.557684e+00  8.305053 12.190278
#> [165,] 3.629911e-01 9.782890e-03 3.128993e+00 1.253780e-02  7.842544 11.088898
#> [166,] 3.392060e-01 2.348553e-01 3.289389e+00 4.544126e-01  7.542687 11.095820
#> [167,] 2.501834e-02 7.354155e-01 3.579165e+00 1.759839e-04  7.281995 11.130559
#> [168,] 2.057244e+00 5.721167e-01 3.846730e+00 7.667960e-01  7.037451 12.306077
#> [169,] 9.059119e-01 2.626428e-02 3.798589e+00 1.932905e+00  8.945204 12.285704
#> [170,] 2.589149e-02 5.955224e-01 3.543044e+00 8.945591e-02  5.878013 10.468462
#> [171,] 2.160622e-02 7.654250e-01 3.262914e+00 1.005260e-02  6.003003 10.658736
#> [172,] 7.060567e-03 9.393329e-01 2.965015e+00 5.607558e-01  6.222611 10.388239
#> [173,] 7.358352e-01 1.046921e+00 2.665302e+00 1.710689e-01  6.518106 10.449165
#> [174,] 3.535001e+00 2.474742e+00 2.231223e+00 2.663475e+00  5.938991 13.793320
#> [175,] 2.375734e+00 7.720753e+00 1.552927e+00 1.354226e-01  4.894893 13.774607
#> [176,] 6.048831e+00 3.207542e+00 1.181453e+00 3.327446e-01  6.149843 16.643413
#> [177,] 8.259966e-02 4.586919e-02 1.144096e+00 8.700391e-03 12.901752 14.838724
#> [178,] 5.945846e-03 1.588522e-01 1.072674e+00 2.457377e-04 12.834044 15.195174
#> [179,] 5.875032e-02 1.219313e-01 1.012545e+00 1.381380e+00 13.245706 15.835335
#> [180,] 6.595083e-01 3.753102e-02 9.819511e-01 1.755840e+00 13.833960 15.859853
#> [181,] 4.000155e-02 1.072596e-01 1.046742e+00 4.272761e-01 11.455867 14.370887
#> [182,] 1.209091e+00 2.079466e-01 1.139131e+00 8.186970e-02 12.026674 16.985061
#> [183,] 1.310750e-01 1.351082e+00 1.384030e+00 2.525122e-02 12.492477 16.722832
#> [184,] 1.068428e+00 1.946902e+00 1.710177e+00 7.920043e-04 11.210948 18.549246
#> [185,] 6.153381e-01 4.243870e+00 2.253334e+00 2.900950e-01 10.549800 19.942329
#> [186,] 5.845489e-05 6.579838e+00 3.036407e+00 2.417222e+00  9.474907 20.014831
#> [187,] 1.783939e+00 6.610248e+00 3.946085e+00 1.495981e+00 10.131230 19.797036
#> [188,] 3.930301e-02 2.933565e+00 4.635463e+00 3.926501e+00 11.863803 25.264877
#> [189,] 3.667687e-02 2.515228e+00 5.334168e+00 1.872966e-01  9.805409 18.128784
#> [190,] 5.714037e-01 2.141569e+00 6.035124e+00 1.911573e-02 10.260600 18.130712
#> [191,] 2.909978e+00 9.547133e-01 6.554284e+00 4.156296e-01 10.670322 19.855165
#> [192,] 5.160795e-01 1.493935e-02 6.562843e+00 5.699957e-01 14.066791 20.434791
#> [193,] 3.530534e-01 3.431089e-01 6.352928e+00 3.392866e-02 15.574028 20.457295
#> [194,] 3.256425e+00 9.400531e-01 5.968610e+00 9.468634e-01 14.748734 22.712958
#> [195,] 1.585012e-01 4.559884e+00 5.080151e+00 6.153309e-01 13.812314 22.089083
#> [196,] 3.175537e-01 3.531806e+00 4.359723e+00 4.160473e-02 13.917511 21.346206
#> [197,] 4.318196e-01 5.038780e+00 3.554284e+00 1.592885e+00 12.467100 21.864164
#> [198,] 2.240887e+00 7.136574e+00 2.685917e+00 9.219020e-02 11.781877 21.965527
#> [199,] 1.398501e-01 2.904812e+00 2.204809e+00 5.209583e+00 12.274627 18.655859
#> [200,] 5.077759e+00 2.142506e+00 1.831099e+00 3.901673e+00 11.880557 17.569540
#> [201,] 2.269323e+00 8.570851e+00 1.152280e+00 8.628903e-03  4.962469 16.773854
#> [202,] 1.142807e-01 3.814682e+00 7.881355e-01 5.572690e-01  9.255665 14.846845
#> [203,] 6.840029e-01 3.015034e+00 5.200504e-01 7.563696e-02  9.034663 13.012851
#> [204,] 1.722498e-01 1.441151e+00 3.682539e-01 7.580750e-01  6.929186  9.217834
#> [205,] 2.303285e-03 8.680796e-01 2.687853e-01 8.894880e-01  5.713735  6.610611
#> [206,] 1.445505e+00 9.281100e-01 1.800304e-01 1.163080e+00  5.708308  6.609896
#> [207,] 6.002105e-02 3.192099e-02 1.699839e-01 1.236822e-02  4.128361  4.587970
#> [208,] 3.740160e-02 3.503620e-04 1.745915e-01 8.477315e-02  4.372803  4.630188
#> [209,] 3.258302e-02 2.104456e-02 1.679192e-01 5.537371e-01  4.446332  4.626836
#> [210,] 2.415023e-02 7.393410e-04 1.726074e-01 5.510238e-01  4.358587  4.658810
#> [211,] 9.160129e-02 1.661050e-02 1.681974e-01 2.703632e-02  4.201448  4.484490
#> [212,] 8.335374e-01 4.775482e-03 1.843638e-01 2.329552e+00  4.122036  7.829061
#> [213,] 3.711173e+00 2.818107e-01 1.421856e-01 1.394843e+00  7.188804  9.953589
#> [214,] 2.440471e-02 3.275153e+00 2.608291e-02 8.478283e-02  4.717387  8.586537
#> [215,] 1.555245e+00 2.988666e+00 3.330764e-03 1.180837e-02  4.724267  8.377537
#> [216,] 1.623039e+00 8.617367e-01 3.503201e-02 1.938907e-01  6.194924  7.694509
#> [217,] 8.246836e-03 3.724589e-03 4.556803e-02 4.839205e-02  6.786901  6.820595
#> [218,] 5.247588e-02 1.527794e-04 5.641032e-02 3.571562e-02  6.806319  6.806319
#> [219,] 2.539318e-01 1.921084e-02 8.324929e-02 2.708127e-01  6.795568  6.795568
#> [220,] 6.392288e-02 1.776910e-01 1.650255e-01 3.206891e-03  6.651936  6.651936
#> [221,] 2.912601e-02 8.096637e-02 2.802386e-01 1.878513e-01  6.494215  6.494215
#> [222,] 2.115945e-01 3.676286e-02 4.521565e-01 1.012297e-01  6.435099  6.435099
#> [223,] 6.886867e-06 2.182020e-01 9.222558e-01 1.688014e+00  6.365863  6.365863
#> [224,] 3.602698e+00 2.324573e-01 1.978091e+00 3.529255e-01  6.352240  6.352240
#> [225,] 1.082933e+00 4.933088e-01 2.575446e+00 5.661918e-01  3.267854  3.267854
#> [226,] 3.959411e-03 2.164581e+00 2.189352e+00 5.135684e-01  2.759488  2.759488
#> [227,] 1.282395e+00 2.591206e+00 1.014522e+00 7.860629e-03  2.597233  2.597233
#> [228,] 1.014522e+00 1.014522e+00          NaN 1.014522e+00  1.014522  1.014522
#> 
#> 
#> 
#> $likelihood
#> $likelihood$initial
#> $likelihood$initial$ll
#> [1] -1233.697
#> 
#> $likelihood$initial$ssq
#> [1] 34176.83
#> 
#> $likelihood$initial$nobs
#> [1] 228
#> 
#> $likelihood$initial$neffectiveobs
#> [1] -1
#> 
#> $likelihood$initial$nparams
#> [1] 4
#> 
#> $likelihood$initial$df
#> [1] 211
#> 
#> $likelihood$initial$aic
#> [1] 2475.393
#> 
#> $likelihood$initial$aicc
#> [1] 2475.584
#> 
#> $likelihood$initial$bic
#> [1] 2488.876
#> 
#> $likelihood$initial$bic2
#> [1] NaN
#> 
#> $likelihood$initial$bicc
#> [1] NaN
#> 
#> $likelihood$initial$hannanquinn
#> [1] 2480.841
#> 
#> 
#> $likelihood$final
#> $likelihood$final$ll
#> [1] -1233.634
#> 
#> $likelihood$final$ssq
#> [1] 58486.03
#> 
#> $likelihood$final$nobs
#> [1] 228
#> 
#> $likelihood$final$neffectiveobs
#> [1] -1
#> 
#> $likelihood$final$nparams
#> [1] 4
#> 
#> $likelihood$final$df
#> [1] 211
#> 
#> $likelihood$final$aic
#> [1] 2475.267
#> 
#> $likelihood$final$aicc
#> [1] 2475.458
#> 
#> $likelihood$final$bic
#> [1] 2488.75
#> 
#> $likelihood$final$bic2
#> [1] NaN
#> 
#> $likelihood$final$bicc
#> [1] NaN
#> 
#> $likelihood$final$hannanquinn
#> [1] 2480.715
#> 
#> 
#> 
#> attr(,"class")
#> [1] "JD3_STS_OUTLIERS"