Title

Title

Usage

sts_outliers(
  y,
  period = NA,
  X = NULL,
  X.td = NULL,
  level = 1,
  slope = 1,
  noise = 1,
  seasonal = c("Trigonometric", "Dummy", "Crude", "HarrisonStevens", "Fixed", "Unused"),
  ao = TRUE,
  ls = TRUE,
  so = FALSE,
  cv = 0,
  tcv = 0,
  estimation.forward = c("Score", "Point", "Full"),
  estimation.backward = c("Point", "Score", "Full")
)

Arguments

y

input time series.

period

annual frequency.

X

Regression variables (same length as y) or NULL

X.td

Groups of days for trading days regressors. The length of the array must be 7. It indicates to what group each week day belongs. The first item corresponds to Mondays and the last one to Sundays. The group used for contrasts (usually Sundays) is identified by 0. The other groups are identified by 1, 2,... n (<= 6). For instance, usual trading days are defined by c(1,2,3,4,5,6,0), week days by c(1,1,1,1,1,0,0), etc...

level

-1 = no level, 0 = fixed level, 1 = sotchastic level

seasonal

Seasonal model

ao, ls, so

boolean indicating if additive outliers (ao), level shift (ls) and seasonal outliers (so) should be detected.

estimation.backward

Examples

 x<-rjd3toolkit::Retail$BookStores
 sts_outliers(x)
#> $model
#> $model$y
#>   [1]  790  540  536  524  553  589  593  895  863  647  642 1166  999  568  602
#>  [16]  583  613  619  608  985  905  669  693 1275 1055  636  635  610  684  726
#>  [31]  679 1156 1023  733  772 1410 1309  720  696  689  786  808  783 1248 1103
#>  [46]  747  851 1468 1375  758  743  751  855  839  787 1361 1042  899  908 1600
#>  [61] 1559  825  813  806  893  876  830 1301 1159  926  996 1771 1467  923  883
#>  [76]  854  920  939  882 1385 1247  954  991 1849 1514  990  976  934  977 1032
#>  [91] 1024 1388 1242 1020 1090 1998 1506 1091 1031  975 1105 1123 1049 1595 1407
#> [106] 1015 1100 1895 1586 1070 1063  935 1071 1078 1006 1821 1407 1016 1099 1958
#> [121] 1973 1001 1009  968 1118 1046 1038 1799 1491  998 1060 1949 2056  992  926
#> [136]  976 1109 1182 1130 2142 1574 1047 1037 2065 2148 1108 1073 1023 1112 1206
#> [151] 1172 2119 1560 1069 1088 2199 2075 1086 1101  997 1120 1154 1214 2219 1553
#> [166] 1064 1150 2256 2225 1101 1092  987 1176 1238 1135 2142 1555 1055 1143 2134
#> [181] 2222 1040 1029  931 1135 1162 1221 2330 1579 1129 1227 2179 2301 1157 1026
#> [196] 1005 1159 1077 1124 2435 1481 1032 1044 2029 2214  985  961  929 1071 1058
#> [211] 1056 2337 1548  975  979 1938 2178  962  962  871 1032 1040 1025 2183 1448
#> [226]  949 1030 1982
#> 
#> $model$variables
#> NULL
#> 
#> $model$X
#> NULL
#> 
#> $model$b
#> numeric(0)
#> 
#> $model$bcov
#> <0 x 0 matrix>
#> 
#> $model$components
#>                 [,1]      [,2]          [,3]       [,4]
#>   [1,]  -7.844922961  654.7048  6.0703506710  143.14015
#>   [2,]   2.583070370  664.4507  6.0668992162 -127.03377
#>   [3,]   0.959926877  672.9829  6.0611327552 -137.94286
#>   [4,]   0.492516282  681.0596  6.0534736176 -157.55216
#>   [5,]   0.530827484  688.8979  6.0441384910 -136.42876
#>   [6,]   1.692139409  696.4782  6.0333609185 -109.17033
#>   [7,]   2.913108327  703.2548  6.0218853742 -113.16795
#>   [8,]  -2.492002974  708.6551  6.0109935103  188.83686
#>   [9,]   0.758725707  715.2121  5.9995889442  147.02914
#>  [10,]   2.551131729  721.4022  5.9880054852  -76.95337
#>  [11,]   0.334115637  726.3855  5.9773655300  -84.71958
#>  [12,]  -3.121336822  731.2015  5.9678160762  437.91982
#>  [13,]   6.535554483  737.4705  5.9579838596  254.99399
#>  [14,]  -2.301302753  740.6675  5.9507442646 -170.36617
#>  [15,]   1.296335324  744.9355  5.9450848094 -144.23180
#>  [16,]   2.771461815  748.5904  5.9415758301 -168.36189
#>  [17,]  -0.211586678  750.9434  5.9414366599 -137.73179
#>  [18,]  -2.817457682  753.3953  5.9445742090 -131.57786
#>  [19,]  -1.635246693  757.1705  5.9497489080 -147.53521
#>  [20,]  -2.636196503  761.7169  5.9562413131  225.91926
#>  [21,]  -1.968861497  767.5050  5.9628916025  139.46382
#>  [22,]   0.035353721  774.2223  5.9688335554 -105.25762
#>  [23,]   1.204789966  780.9289  5.9740827258  -89.13366
#>  [24,]   0.592373915  787.0762  5.9791691737  487.33139
#>  [25,]  -8.221872552  792.9512  5.9843535200  270.27072
#>  [26,]   0.630229266  802.6834  5.9860184672 -167.31367
#>  [27,]   0.640036579  812.1221  5.9844412910 -177.76215
#>  [28,]  -2.359443989  821.2593  5.9799035822 -208.89989
#>  [29,]  -0.682286472  831.4975  5.9713672790 -146.81519
#>  [30,]   1.188165100  842.0468  5.9585322026 -117.23493
#>  [31,]  -1.882222698  852.0265  5.9419210983 -171.14430
#>  [32,]   3.291636898  862.8715  5.9207058630  289.83682
#>  [33,]   1.531696631  872.1531  5.8963346861  149.31518
#>  [34,]   0.328635278  880.6927  5.8694814535 -148.02133
#>  [35,]  -2.279008595  889.0514  5.8402907515 -114.77242
#>  [36,]   1.907170459  898.4488  5.8077599069  509.64408
#>  [37,]   7.644355474  906.9200  5.7727279990  394.43566
#>  [38,]   0.390767602  911.7746  5.7385582408 -192.16535
#>  [39,]  -2.259304900  916.4119  5.7054225545 -218.15261
#>  [40,]  -0.733021868  922.0747  5.6723269356 -232.34165
#>  [41,]  -0.541906935  928.0478  5.6389488837 -141.50586
#>  [42,]   2.370749390  934.2414  5.6050499811 -128.61214
#>  [43,]   3.724334720  939.2903  5.5716732635 -160.01468
#>  [44,]  -0.313161629  942.5610  5.5404572903  305.75219
#>  [45,]   6.863763116  945.9471  5.5112642830  150.18914
#>  [46,]  -7.271686286  946.0882  5.4871140247 -191.81647
#>  [47,]   1.017953919  949.6121  5.4648072623  -99.63002
#>  [48,]  -3.337371551  952.6367  5.4447918549  518.70064
#>  [49,]  -4.802770425  957.2050  5.4255994916  422.59774
#>  [50,]  -0.934386402  964.0044  5.4051171440 -205.07000
#>  [51,]  -0.890245085  971.2210  5.3829337191 -227.33079
#>  [52,]   0.527086456  978.8326  5.3586575445 -228.35970
#>  [53,]   2.434999937  986.1730  5.3325205173 -133.60796
#>  [54,]  -0.984796045  992.3463  5.3055939409 -152.36150
#>  [55,]  -2.001160155  998.9541  5.2774445439 -209.95296
#>  [56,]   9.130006676 1006.4714  5.2471918955  345.39860
#>  [57,] -10.686128905 1009.6807  5.2188528361   43.00539
#>  [58,]   6.046803956 1017.8685  5.1877258932 -124.91531
#>  [59,]  -2.255921477 1023.1920  5.1564714205 -112.93612
#>  [60,]  -3.663629242 1029.5413  5.1240969021  574.12234
#>  [61,]  15.036750612 1037.5747  5.0889904866  506.38857
#>  [62,]  -2.868018021 1038.5278  5.0577677474 -210.65979
#>  [63,]  -1.494645191 1040.7935  5.0291668704 -226.29881
#>  [64,]   0.142461827 1043.7308  5.0025302711 -237.87326
#>  [65,]  -0.099130953 1046.5747  4.9779206270 -153.47562
#>  [66,]  -1.435968641 1049.4405  4.9552943244 -172.00457
#>  [67,]  -1.458629328 1052.9565  4.9340195941 -221.49787
#>  [68,]  -7.384227970 1057.1346  4.9134546971  251.24964
#>  [69,]   1.282419123 1064.7518  4.8903508662   92.96574
#>  [70,]   0.633156364 1071.7451  4.8652723149 -146.37830
#>  [71,]   5.357322210 1078.4167  4.8384976068  -87.77404
#>  [72,]   5.607470024 1082.5514  4.8123837512  682.84108
#>  [73,]  -6.665940650 1084.0328  4.7893978122  389.63314
#>  [74,]   2.019022645 1088.6144  4.7666070405 -167.63338
#>  [75,]  -0.248521523 1092.2272  4.7448997260 -208.97863
#>  [76,]  -1.453401978 1095.9347  4.7241665290 -240.48128
#>  [77,]  -1.140656386 1100.3024  4.7037680104 -179.16178
#>  [78,]  -1.783909021 1105.1842  4.6832023266 -164.40032
#>  [79,]  -4.588095918 1110.8813  4.6616846286 -224.29317
#>  [80,]   3.510925249 1118.7064  4.6371963360  262.78264
#>  [81,]   5.671920161 1124.8622  4.6112821168  116.46593
#>  [82,]  -1.562512196 1128.3345  4.5864373907 -172.77198
#>  [83,]  -5.070629408 1132.5141  4.5619747143 -136.44344
#>  [84,]  -2.034907140 1139.0449  4.5356632120  711.98998
#>  [85,]   2.865531575 1146.5029  4.5066076053  364.63158
#>  [86,]  -0.854209710 1152.5892  4.4760686144 -161.73499
#>  [87,]   2.723439439 1159.0452  4.4436704210 -185.76864
#>  [88,]   3.007687552 1164.1928  4.4106112303 -233.20048
#>  [89,]  -3.610844790 1167.8981  4.3782143055 -187.28729
#>  [90,]   0.632536833 1173.2629  4.3448910181 -141.89540
#>  [91,]   7.711548075 1178.2979  4.3109196593 -162.00946
#>  [92,] -13.211736251 1179.6859  4.2796930042  221.52584
#>  [93,]  -9.676182705 1187.2327  4.2453984131   64.44344
#>  [94,]   2.370038312 1199.2789  4.2037787498 -181.64891
#>  [95,]   3.902442692 1210.1729  4.1558767370 -124.07539
#>  [96,]  11.502967982 1219.1907  4.1034092946  767.30632
#>  [97,]  -9.023230135 1222.7665  4.0514372720  292.25670
#>  [98,]   4.005500026 1230.5180  3.9959908061 -143.52353
#>  [99,]  -0.512418095 1236.3374  3.9388321578 -204.82497
#> [100,]   1.150846166 1242.3397  3.8797358836 -268.49051
#> [101,]   6.671673912 1247.7436  3.8192083100 -149.41531
#> [102,]   5.172955477 1249.9612  3.7601847090 -132.13417
#> [103,]   1.914723869 1249.6961  3.7049409754 -202.61081
#> [104,]  -3.372502425 1248.4786  3.6543195119  349.89390
#> [105,]   7.885581497 1248.7906  3.6068365511  150.32380
#> [106,]  -0.661428692 1245.3605  3.5659614361 -229.69910
#> [107,]   1.714034886 1242.1995  3.5314031644 -143.91349
#> [108,]  -8.225122982 1238.2008  3.5039158434  665.02437
#> [109,] -16.923034202 1238.0283  3.4798807453  364.89476
#> [110,]   4.280298556 1245.7607  3.4518524069 -180.04100
#> [111,]   7.193252118 1251.4596  3.4217139895 -195.65290
#> [112,]  -3.741625275 1253.7582  3.3926302386 -315.01659
#> [113,]  -3.024915213 1257.7808  3.3629549877 -183.75584
#> [114,]  -1.336520362 1263.1909  3.3313573940 -183.85436
#> [115,]  -4.563623626 1269.1956  3.2972494417 -258.63198
#> [116,]  12.762181031 1277.3044  3.2586233174  530.93340
#> [117,]  -5.477054614 1279.3951  3.2210938744  133.08191
#> [118,]  -0.524068633 1284.0145  3.1822514259 -267.49044
#> [119,]   0.427978973 1288.8406  3.1418654026 -190.26854
#> [120,]   2.021188611 1293.4257  3.1001240983  662.55309
#> [121,]  15.320870002 1297.0221  3.0579167556  660.65699
#> [122,]  -4.172300284 1293.3981  3.0219839494 -288.22578
#> [123,]  -0.596785738 1291.6929  2.9904900330 -282.09615
#> [124,]   3.073662401 1290.2359  2.9631724445 -325.30957
#> [125,]   2.956784803 1287.3115  2.9413834733 -172.26824
#> [126,]  -5.988147793 1282.9799  2.9264239874 -230.99173
#> [127,]  -2.389320028 1281.4390  2.9156594395 -241.04964
#> [128,] -14.896048004 1281.0067  2.9080386234  532.88930
#> [129,]   2.402285283 1287.5462  2.8970078694  201.05156
#> [130,]   0.927469265 1292.9490  2.8836240876 -295.87645
#> [131,]   1.075606884 1297.9039  2.8682953277 -238.97949
#> [132,]  -3.691217088 1302.3395  2.8514948140  650.35172
#> [133,]   4.057912213 1308.4878  2.8315985574  743.45433
#> [134,]  -4.693093695 1312.7149  2.8103918784 -316.02178
#> [135,] -11.968832535 1319.1196  2.7858100020 -381.15080
#> [136,]  -1.346180431 1331.1076  2.7525871168 -353.76138
#> [137,]  -0.382656446 1343.6930  2.7101309569 -234.31034
#> [138,]   4.510319455 1356.4153  2.6582731687 -178.92557
#> [139,]   1.076687998 1366.9724  2.5989981138 -238.04913
#> [140,]  17.176285300 1376.9659  2.5327794920  747.85783
#> [141,]   3.556400691 1378.8455  2.4671741872  191.59807
#> [142,]  -2.467539790 1378.9933  2.4037468746 -329.52574
#> [143,]  -7.387587747 1380.2337  2.3414119346 -335.84613
#> [144,]  -5.195802768 1384.8731  2.2769191218  685.32268
#> [145,]   6.556456317 1391.8824  2.2079824892  749.56112
#> [146,]   4.011785861 1395.7509  2.1374866173 -291.76268
#> [147,]   3.793438210 1397.6692  2.0671965328 -328.46267
#> [148,]   1.148142252 1397.7399  1.9987811978 -375.88809
#> [149,]  -2.379931636 1397.2043  1.9327457494 -282.82437
#> [150,]   2.847384779 1397.7177  1.8680431121 -194.56507
#> [151,]   0.312619517 1396.8323  1.8059260222 -225.14491
#> [152,]  -6.134270683 1395.7383  1.7465320199  729.39597
#> [153,]   0.334910527 1397.4590  1.6871622696  162.20608
#> [154,]   1.441749445 1398.9634  1.6279641186 -331.40518
#> [155,]  -0.001036659 1399.7331  1.5695718794 -311.73211
#> [156,]   5.603357779 1400.4450  1.5119850959  792.95168
#> [157,] -11.515928192 1398.4738  1.4576690232  688.04208
#> [158,]  -1.225861243 1401.8440  1.4015571100 -314.61811
#> [159,]   3.924971457 1405.7323  1.3431100258 -308.65731
#> [160,]  -0.685785526 1407.7233  1.2840546025 -410.03750
#> [161,]  -2.755077626 1409.9765  1.2240891216 -287.22142
#> [162,]  -8.638023264 1413.4606  1.1620014582 -250.82255
#> [163,]   5.511641594 1420.9297  1.0939912250 -212.44137
#> [164,]   7.137177661 1425.7485  1.0224833255  786.11431
#> [165,]  -3.169159112 1427.1518  0.9506178346  129.01735
#> [166,]  -3.063569624 1429.9681  0.8770004482 -362.90451
#> [167,]   0.832002946 1434.1461  0.8002833171 -284.97811
#> [168,]   7.544642303 1437.8576  0.7208324896  810.59776
#> [169,]   5.006549460 1437.9548  0.6419673089  782.03869
#> [170,]   0.846396207 1435.6273  0.5658904581 -335.47373
#> [171,]   0.773186944 1432.8273  0.4929743179 -341.60046
#> [172,]   0.441992061 1429.5920  0.4235590598 -443.03402
#> [173,]   4.512167281 1426.0803  0.3578391429 -254.59246
#> [174,]   9.889845672 1420.3887  0.2977997416 -192.27859
#> [175,]  -8.107620458 1410.0035  0.2477919955 -266.89586
#> [176,] -12.936893382 1403.3669  0.2042488734  751.57003
#> [177,]   1.511761697 1402.7480  0.1614786557  150.74022
#> [178,]  -0.405602059 1401.3781  0.1201464575 -345.97250
#> [179,]  -1.274965667 1400.1569  0.0800738300 -255.88192
#> [180,]  -4.271727447 1399.4930  0.0406998389  738.77877
#> [181,]  -1.052038070 1400.7911  0.0001450937  822.26096
#> [182,]  -5.783924301 1402.5416 -0.0420532606 -356.75764
#> [183,]  -1.904376355 1406.9598 -0.0884399211 -376.05541
#> [184,]  -5.437071711 1412.2239 -0.1398527379 -475.78681
#> [185,]  -4.126184550 1419.9840 -0.1986838067 -280.85780
#> [186,]  -0.040216374 1429.6185 -0.2667484850 -267.57829
#> [187,]   7.025547025 1439.2038 -0.3440644663 -225.22934
#> [188,]   1.042802563 1445.4201 -0.4275407888  883.53710
#> [189,]   1.007359071 1451.0643 -0.5167186608  126.92830
#> [190,]   3.976119789 1456.1474 -0.6111548843 -331.12355
#> [191,]   8.972892488 1459.2732 -0.7091001243 -241.24604
#> [192,]   3.778729382 1458.0969 -0.8066066643  717.12441
#> [193,]   3.125389957 1455.0526 -0.9020120134  842.82199
#> [194,]   9.491933325 1450.4486 -0.9939411239 -302.94058
#> [195,]  -2.094111182 1441.3055 -1.0782179221 -413.21137
#> [196,]   2.964087678 1433.0592 -1.1557637336 -431.02328
#> [197,]   3.456429746 1423.3466 -1.2252744776 -267.80303
#> [198,]  -7.873786900 1411.9451 -1.2852294621 -327.07127
#> [199,]  -1.966938873 1404.1727 -1.3390928408 -278.20572
#> [200,]  11.851951700 1397.2680 -1.3877299889 1025.88008
#> [201,]  -7.922863662 1384.7617 -1.4259265180  104.16121
#> [202,]  -1.777932430 1375.9292 -1.4571681735 -342.15130
#> [203,]  -4.349553760 1367.8986 -1.4822371753 -319.54903
#> [204,]  -2.182704284 1361.8808 -1.5030471543  669.30194
#> [205,]   0.252002924 1356.8648 -1.5205584138  856.88320
#> [206,]  -6.312910015 1351.7132 -1.5346600825 -360.40033
#> [207,]  -1.286376123 1349.5054 -1.5481295903 -387.21899
#> [208,]   1.015455810 1347.8867 -1.5615328916 -419.90219
#> [209,]  -0.947565342 1345.7789 -1.5744232261 -273.83136
#> [210,]   0.815768210 1344.1022 -1.5872174850 -286.91796
#> [211,]  -1.587235121 1342.0304 -1.5995567627 -284.44321
#> [212,]   4.787221983 1340.6900 -1.6121393796  991.52275
#> [213,]  10.072833143 1337.0941 -1.6228591480  200.83309
#> [214,]  -0.816190373 1328.7680 -1.6272844212 -352.95179
#> [215,]  -6.466255664 1320.8199 -1.6257742905 -335.35362
#> [216,]  -6.584901344 1315.9029 -1.6211736517  628.68199
#> [217,]  -0.351635828 1314.0758 -1.6163796001  864.27587
#> [218,]   0.887002566 1312.4182 -1.6115468416 -351.30517
#> [219,]   1.951029778 1310.3498 -1.6062851301 -350.30084
#> [220,]  -0.978869523 1307.3726 -1.5997360901 -435.39373
#> [221,]  -0.660140137 1304.8606 -1.5923303852 -272.20043
#> [222,]   1.778961759 1302.6652 -1.5843584509 -264.44421
#> [223,]  -0.010118228 1299.6444 -1.5750376147 -274.63428
#> [224,]  -7.313149239 1296.6376 -1.5643723282  893.67555
#> [225,]  -3.979406154 1297.0679 -1.5555800867  154.91150
#> [226,]   0.240072224 1299.3715 -1.5504116711 -350.61154
#> [227,]   4.245818375 1301.5677 -1.5487614591 -275.81353
#> [228,]   3.750820212 1301.7763 -1.5487614591  676.47286
#> 
#> $model$linearized
#>   [1]  790  540  536  524  553  589  593  895  863  647  642 1166  999  568  602
#>  [16]  583  613  619  608  985  905  669  693 1275 1055  636  635  610  684  726
#>  [31]  679 1156 1023  733  772 1410 1309  720  696  689  786  808  783 1248 1103
#>  [46]  747  851 1468 1375  758  743  751  855  839  787 1361 1042  899  908 1600
#>  [61] 1559  825  813  806  893  876  830 1301 1159  926  996 1771 1467  923  883
#>  [76]  854  920  939  882 1385 1247  954  991 1849 1514  990  976  934  977 1032
#>  [91] 1024 1388 1242 1020 1090 1998 1506 1091 1031  975 1105 1123 1049 1595 1407
#> [106] 1015 1100 1895 1586 1070 1063  935 1071 1078 1006 1821 1407 1016 1099 1958
#> [121] 1973 1001 1009  968 1118 1046 1038 1799 1491  998 1060 1949 2056  992  926
#> [136]  976 1109 1182 1130 2142 1574 1047 1037 2065 2148 1108 1073 1023 1112 1206
#> [151] 1172 2119 1560 1069 1088 2199 2075 1086 1101  997 1120 1154 1214 2219 1553
#> [166] 1064 1150 2256 2225 1101 1092  987 1176 1238 1135 2142 1555 1055 1143 2134
#> [181] 2222 1040 1029  931 1135 1162 1221 2330 1579 1129 1227 2179 2301 1157 1026
#> [196] 1005 1159 1077 1124 2435 1481 1032 1044 2029 2214  985  961  929 1071 1058
#> [211] 1056 2337 1548  975  979 1938 2178  962  962  871 1032 1040 1025 2183 1448
#> [226]  949 1030 1982
#> 
#> 
#> $bsm
#> $bsm$initial
#> $bsm$initial$level
#> [1] 0.8093253
#> 
#> $bsm$initial$slope
#> [1] 0.0006285975
#> 
#> $bsm$initial$seasonal
#> [1] 0.3366748
#> 
#> $bsm$initial$noise
#> [1] 1
#> 
#> $bsm$initial$tau
#>                [,1]         [,2]         [,3]         [,4]      [,5]      [,6]
#>   [1,] 4.403298e+00          NaN          NaN          NaN  0.000000  4.403298
#>   [2,] 4.565711e-01 4.403298e+00 4.403298e+00          NaN  4.403298  4.600546
#>   [3,] 5.520470e-02 1.125366e+00 2.710499e+00          NaN  4.600546  4.612089
#>   [4,] 9.446070e-03 6.098100e-01 1.959419e+00          NaN  4.612089  4.613230
#>   [5,] 1.751139e-02 4.345789e-01 1.542688e+00          NaN  4.613230  4.626308
#>   [6,] 1.939122e-01 3.023004e-01 1.270862e+00          NaN  4.626308  4.802833
#>   [7,] 5.466269e-01 7.488327e-02 9.770262e-01          NaN  4.802833  5.201798
#>   [8,] 3.802374e-01 2.394507e-02 6.462570e-01          NaN  5.201798  5.452091
#>   [9,] 3.748188e-02 3.735837e-02 5.369590e-01          NaN  5.452091  5.487848
#>  [10,] 4.075649e-01 6.989694e-03 4.367873e-01          NaN  5.487848  5.960464
#>  [11,] 3.834970e-03 6.991890e-02 3.038990e-01          NaN  5.960464  6.002424
#>  [12,] 6.308686e-01 8.654182e-02 2.095099e-01 4.403298e+00  6.002424  6.424913
#>  [13,] 1.743675e+00 1.689755e-02 1.886464e-01 3.450110e+00  6.424529  6.428569
#>  [14,] 2.020486e-01 6.011080e-01 9.557011e-02 8.604649e-02  2.479746  6.428861
#>  [15,] 5.203938e-02 2.063477e-01 5.675891e-02 8.785514e-03  2.165904  3.413090
#>  [16,] 2.906174e-01 3.589313e-01 2.449310e-02 6.075287e-04  2.286885  3.630281
#>  [17,] 1.948972e-03 8.963982e-01 1.522328e-03 4.474846e-02  2.671009  3.562634
#>  [18,] 2.946655e-01 8.369839e-01 4.671676e-03 4.271627e-02  2.670483  3.386514
#>  [19,] 9.931803e-02 3.117365e-01 1.696004e-02 8.698266e-01  2.768753  3.236133
#>  [20,] 2.472950e-01 1.239901e-01 2.775968e-02 3.106652e-01  2.419538  4.020427
#>  [21,] 1.439755e-01 7.515383e-04 2.764578e-02 9.369706e-02  3.764691  3.883878
#>  [22,] 2.799738e-05 4.847032e-02 1.947478e-02 1.573756e-01  3.785125  3.992229
#>  [23,] 6.087396e-02 4.997264e-02 1.289342e-02 3.471622e-01  3.673072  3.906779
#>  [24,] 1.411248e-02 3.918681e-03 1.106264e-02 1.469284e-01  3.865308  4.272494
#>  [25,] 2.563692e+00 2.209079e-04 1.103712e-02 3.773772e-02  3.684232  9.953356
#>  [26,] 1.576644e-02 1.057209e+00 2.427887e-12 1.168561e-01  9.058751  9.984942
#>  [27,] 1.883723e-02 8.928815e-01 8.855653e-03 3.959876e-02  8.973490 10.036564
#>  [28,] 2.110754e-01 7.300257e-01 3.140372e-02 2.738798e-01  8.778765  9.867598
#>  [29,] 1.369133e-02 1.325468e+00 8.277611e-02 6.703082e-02  7.972566  9.732651
#>  [30,] 4.771731e-02 1.502704e+00 1.628070e-01 1.766659e-01  7.909383  9.726674
#>  [31,] 1.287965e-01 1.173490e+00 2.538015e-01 1.215772e+00  7.976144  9.593135
#>  [32,] 3.891283e-01 1.728113e+00 3.902797e-01 4.454588e-01  7.808803  9.918874
#>  [33,] 7.247109e-02 8.289263e-01 4.982741e-01 4.336672e-01  6.811803  8.126507
#>  [34,] 7.540653e-03 5.417123e-01 5.925940e-01 2.509829e-02  6.779255  8.174214
#>  [35,] 1.966725e-01 4.617821e-01 6.854140e-01 5.027248e-01  7.073062  8.069729
#>  [36,] 1.355323e-01 9.320100e-01 8.318988e-01 1.539040e+00  6.980967  8.084535
#>  [37,] 2.217785e+00 5.291978e-01 9.479907e-01 3.328123e+00  6.754939  7.798301
#>  [38,] 5.606814e-03 5.430131e-02 8.963102e-01 1.245137e-01  1.534149  2.877455
#>  [39,] 1.880161e-01 7.904779e-02 8.391198e-01 1.327104e-01  1.548664  2.668644
#>  [40,] 1.566465e-02 3.068446e-06 8.305217e-01 4.002893e-02  1.108363  1.956431
#>  [41,] 1.419459e-02 6.212215e-03 8.358271e-01 3.339714e-03  1.072767  1.996499
#>  [42,] 2.087788e-01 2.416239e-02 8.542987e-01 1.555789e-02  1.121517  2.330078
#>  [43,] 5.072279e-01 1.918319e-02 8.246019e-01 3.057798e-02  1.568281  2.879736
#>  [44,] 1.057323e-02 3.556205e-01 7.245570e-01 1.106561e-01  1.931138  3.054617
#>  [45,] 1.768649e+00 2.809400e-01 6.413590e-01 2.365425e-01  2.028591  6.908702
#>  [46,] 1.922494e+00 1.916265e+00 4.555796e-01 4.344090e-01  4.867212  7.928142
#>  [47,] 4.475132e-02 2.433616e-01 3.957116e-01 3.275629e-03  6.254159  7.083449
#>  [48,] 3.733384e-01 3.956415e-01 3.266338e-01 3.318233e-02  6.173217  8.154263
#>  [49,] 1.010661e+00 5.598482e-02 3.014360e-01 2.391402e-01  7.539126 10.952040
#>  [50,] 2.564356e-02 1.669505e-01 3.405530e-01 1.618037e-02 10.179463 11.079507
#>  [51,] 1.955488e-02 2.613553e-01 3.931851e-01 5.341313e-03 10.027708 11.018833
#>  [52,] 1.019924e-02 3.609594e-01 4.600424e-01 3.778797e-02  9.834822 10.947578
#>  [53,] 2.228759e-01 2.871869e-01 5.237042e-01 3.619935e-01  9.824493 10.963924
#>  [54,] 3.893396e-02 5.433948e-02 5.513671e-01 1.352078e-02 10.003260 10.963452
#>  [55,] 1.403580e-01 1.292681e-01 5.966756e-01 8.917814e-01  9.790647 10.753746
#>  [56,] 3.041641e+00 3.594789e-01 6.777054e-01 2.840282e+00  8.960284 13.083677
#>  [57,] 4.174672e+00 2.683566e-01 6.033835e-01 4.319256e+00 11.990886 13.408271
#>  [58,] 1.350306e+00 6.257019e-01 7.126243e-01 7.219571e-01  9.785786 11.032441
#>  [59,] 2.036488e-01 2.168341e-03 7.172594e-01 5.174773e-01  9.076376  9.788033
#>  [60,] 4.971617e-01 1.126079e-01 7.651623e-01 3.073902e-01  9.322652 12.821821
#>  [61,] 8.290068e+00 6.194103e-01 8.865898e-01 1.656289e+00 11.224992 16.490118
#>  [62,] 2.967641e-01 1.116268e+00 7.210289e-01 1.167244e-02 11.702808 14.132154
#>  [63,] 8.112893e-02 5.007793e-01 6.194574e-01 2.380470e-02 12.455919 13.988171
#>  [64,] 1.206892e-03 2.758894e-01 5.489175e-01 2.432880e-02 12.699906 13.919847
#>  [65,] 1.054841e-03 2.996740e-01 4.801917e-01 9.209752e-04 12.661137 13.760033
#>  [66,] 6.274799e-02 2.772866e-01 4.185542e-01 2.615197e-03 12.611050 13.420539
#>  [67,] 6.386095e-02 1.341301e-01 3.780529e-01 1.115009e+00 12.507901 13.080842
#>  [68,] 2.075009e+00 4.182632e-02 3.560988e-01 3.573187e+00 12.438328 13.096117
#>  [69,] 6.827552e-02 5.141900e-01 4.326504e-01 8.983986e-01  9.631445 10.691937
#>  [70,] 1.054805e-02 3.023412e-01 4.961515e-01 4.535624e-01  9.262236 10.123886
#>  [71,] 1.087743e+00 2.343679e-01 5.555139e-01 9.196772e-01  9.539258 10.837332
#>  [72,] 1.166788e+00 3.348631e-02 5.316192e-01 8.006795e+00 10.317290 10.794763
#>  [73,] 1.780407e+00 7.635455e-01 4.275126e-01 3.785789e+00  6.429268  7.773240
#>  [74,] 1.546437e-01 4.178766e-04 4.246843e-01 1.886929e-02  2.419141  2.720019
#>  [75,] 2.156063e-03 7.393459e-02 3.944341e-01 2.079381e-01  1.961885  2.618019
#>  [76,] 8.746197e-02 5.866153e-02 3.684340e-01 8.558896e-03  1.999917  2.636333
#>  [77,] 3.829384e-02 2.819036e-03 3.625575e-01 5.578215e-02  2.036648  2.509854
#>  [78,] 1.162363e-01 5.187188e-03 3.696288e-01 2.833137e-03  1.912763  2.667302
#>  [79,] 8.009272e-01 8.409095e-02 3.999618e-01 4.286555e-03  2.134807  4.819665
#>  [80,] 5.689483e-01 7.432124e-01 4.979761e-01 8.042534e-02  3.556227  5.122354
#>  [81,] 1.227781e+00 1.443040e-01 5.442963e-01 6.369771e-01  4.296275  5.599033
#>  [82,] 9.531952e-02 1.078751e-01 5.033551e-01 4.589748e-02  4.777617  5.397711
#>  [83,] 1.004039e+00 1.718275e-02 4.872588e-01 8.534946e-01  4.542677  5.423934
#>  [84,] 2.071866e-01 2.594782e-01 5.493486e-01 3.025483e-01  3.915136  6.050738
#>  [85,] 3.773018e-01 6.404801e-01 6.546118e-01 9.901995e-02  4.566903  6.720602
#>  [86,] 3.282109e-02 1.662077e-01 7.115471e-01 7.714379e-02  5.984608  6.954602
#>  [87,] 2.630908e-01 2.740079e-01 7.882474e-01 4.867599e-02  6.050201  6.965046
#>  [88,] 3.325052e-01 3.827345e-02 8.176358e-01 3.495979e-02  5.956968  6.898861
#>  [89,] 4.965206e-01 2.988759e-02 7.908368e-01 1.393626e-01  6.192055  7.583336
#>  [90,] 1.205282e-02 7.696758e-02 8.329609e-01 3.321012e-01  6.750822  7.739020
#>  [91,] 2.166668e+00 4.295921e-02 8.650507e-01 2.224894e+00  6.674390  8.069208
#>  [92,] 6.184250e+00 5.378136e-01 7.526176e-01 2.540714e-03  5.668288 14.852501
#>  [93,] 3.539069e+00 7.323235e-01 8.837855e-01 2.910801e-01 11.675410 17.970053
#>  [94,] 2.034602e-01 4.234739e+00 1.242793e+00 1.938682e-01 11.673624 16.812014
#>  [95,] 5.672982e-01 3.131427e+00 1.600157e+00 1.661074e-01 12.112139 16.802661
#>  [96,] 4.846994e+00 1.659672e+00 1.888471e+00 2.432713e+00 12.683461 18.962729
#>  [97,] 2.660146e+00 1.403529e-02 1.860341e+00 1.135599e+00 15.669736 19.373813
#>  [98,] 5.289678e-01 8.531448e-01 2.079947e+00 1.425511e-01 14.950248 18.237072
#>  [99,] 2.092307e-02 2.106269e-01 2.193330e+00 7.883489e-02 15.821870 18.179029
#> [100,] 4.611841e-02 3.039927e-01 2.333654e+00 1.045931e+00 15.801217 18.927635
#> [101,] 1.653759e+00 1.715102e-01 2.441794e+00 1.565057e-01 17.033114 19.349152
#> [102,] 9.590122e-01 1.660615e-01 2.334378e+00 5.037893e-01 16.669435 20.312560
#> [103,] 1.169855e-01 1.067467e+00 2.073640e+00 5.044384e+00 17.757653 22.266502
#> [104,] 3.912348e-01 1.566750e+00 1.778592e+00 3.870618e-01 18.847677 29.491316
#> [105,] 2.414297e+00 7.260128e-01 1.590658e+00 2.763808e+00 22.808602 26.781748
#> [106,] 1.558829e-02 3.403187e+00 1.220027e+00 2.404349e-02 19.370570 25.837811
#> [107,] 9.404306e-02 3.115199e+00 9.114198e-01 9.214177e-01 18.925873 24.940450
#> [108,] 2.573450e+00 3.845152e+00 6.212812e-01 4.238341e+00 19.299244 24.021406
#> [109,] 1.028998e+01 8.760969e-01 5.023196e-01 1.347648e+00 17.060943 46.712708
#> [110,] 6.613025e-01 1.239672e+00 6.451401e-01 1.986219e-01 44.025823 47.816319
#> [111,] 1.891017e+00 3.526724e-01 7.285962e-01 3.845167e-01 45.509992 48.733595
#> [112,] 5.293857e-01 8.104613e-02 6.879073e-01 1.283553e-04 47.648892 48.754360
#> [113,] 3.641191e-01 3.245520e-02 7.134797e-01 1.124677e-02 46.838893 48.023990
#> [114,] 6.754901e-02 3.199765e-01 7.968636e-01 1.169160e-02 45.747675 47.456067
#> [115,] 7.206239e-01 5.353983e-01 9.115728e-01 9.981996e+00 45.733829 47.042929
#> [116,] 6.127743e+00 1.623197e+00 1.129721e+00 7.180599e+00 44.475115 46.384671
#> [117,] 1.156918e+00 9.452010e-02 1.074974e+00 6.023499e-03 34.805458 37.336927
#> [118,] 1.309927e-02 1.441986e-01 1.142882e+00 6.367895e-02 35.826080 37.306528
#> [119,] 3.788010e-03 2.050794e-01 1.226597e+00 7.126531e-01 35.618526 37.369368
#> [120,] 1.834764e-01 1.710132e-01 1.305658e+00 1.276436e+01 36.401337 37.397559
#> [121,] 8.591077e+00 1.956611e-02 1.333045e+00 2.457398e+01 36.738021 38.609939
#> [122,] 6.206114e-01 3.002360e+00 1.015202e+00 7.202783e-03 10.193731 15.443014
#> [123,] 1.241645e-02 1.511502e+00 8.159163e-01 1.863628e+00  7.753710 12.839664
#> [124,] 3.731894e-01 1.341539e+00 6.480672e-01 1.121459e-02  7.101754  9.846691
#> [125,] 3.054411e-01 2.398027e+00 4.537803e-01 1.852248e-01  7.419179 11.104860
#> [126,] 1.266546e+00 3.616356e+00 2.623961e-01 2.006647e-01  7.572360 11.253739
#> [127,] 2.011405e-01 1.398668e+00 1.696483e-01 4.348066e-04  8.402405 10.257061
#> [128,] 8.549864e+00 8.030294e-01 1.127861e-01 4.528649e-01  8.805925 26.242603
#> [129,] 2.022785e-01 9.449755e-01 1.750777e-01 9.703365e-01 25.169378 26.417745
#> [130,] 4.395334e-02 4.688797e-01 2.271737e-01 1.751522e-01 24.977336 26.008366
#> [131,] 5.462282e-02 3.033835e-01 2.740248e-01 2.478922e-01 24.585100 26.807722
#> [132,] 4.429482e-01 1.611846e-01 3.109766e-01 1.463740e+00 24.987273 26.739591
#> [133,] 4.501999e-01 6.835010e-01 3.943560e-01 2.091384e+00 25.252257 27.016093
#> [134,] 7.696166e-01 1.584233e-01 4.381139e-01 5.233619e-01 24.463040 27.091791
#> [135,] 5.109517e+00 9.189108e-01 5.527706e-01 4.667430e-01 24.321971 30.293263
#> [136,] 7.673564e-02 5.774510e+00 8.984818e-01 2.593582e-03 20.066347 27.613302
#> [137,] 4.076462e-03 6.656838e+00 1.362872e+00 1.331570e+00 19.630046 27.623595
#> [138,] 7.254201e-01 6.869341e+00 1.933591e+00 3.653572e-02 19.655445 27.533809
#> [139,] 4.751067e-02 4.312788e+00 2.456994e+00 5.560448e+00 19.634170 25.235796
#> [140,] 1.096909e+01 3.753843e+00 3.002188e+00 6.311688e+00 19.610973 24.675098
#> [141,] 4.370668e-01 3.205911e-02 2.951179e+00 5.751236e-01  5.800133  8.950978
#> [142,] 2.415588e-01 3.618817e-01 2.779482e+00 1.946856e-01  5.855783  8.938445
#> [143,] 1.951277e+00 8.263896e-02 2.700099e+00 1.814141e+00  5.497204  8.682121
#> [144,] 1.002313e+00 3.664037e-01 2.873907e+00 2.702344e-03  4.193328 11.312971
#> [145,] 1.589208e+00 1.550591e+00 3.246829e+00 2.426514e-04  6.800881 11.895020
#> [146,] 6.033467e-01 1.931612e-01 3.385542e+00 3.392190e-01  8.314813 11.668814
#> [147,] 5.591178e-01 3.247340e-03 3.370234e+00 1.414679e+00  8.286298 11.679582
#> [148,] 4.916258e-02 2.861893e-01 3.207461e+00 6.359547e-02  6.662544 10.872753
#> [149,] 1.892485e-01 4.579643e-01 3.006663e+00 8.382807e-02  6.778755 10.443699
#> [150,] 3.147619e-01 1.589256e-01 2.892498e+00 1.658058e-01  6.365551 10.360012
#> [151,] 1.269476e-03 5.735882e-01 2.679410e+00 3.170300e-01  7.297201 10.192171
#> [152,] 1.577369e+00 6.086744e-01 2.468307e+00 9.164547e-05  7.227072 11.663803
#> [153,] 4.834416e-03 5.203113e-04 2.477090e+00 1.087850e-01  8.432336 11.316140
#> [154,] 9.521881e-02 4.686187e-04 2.474084e+00 1.710431e-01  8.543783 10.956777
#> [155,] 1.865882e-03 4.794894e-02 2.418461e+00 9.472862e-01  8.728350 11.158775
#> [156,] 1.110198e+00 6.081701e-02 2.356310e+00 4.784714e+00  8.624124 11.348674
#> [157,] 4.891923e+00 8.473186e-01 2.123535e+00 2.231845e-01  5.420046 13.727853
#> [158,] 5.873839e-02 2.441157e-01 2.251871e+00 3.718759e-01  7.762920 11.224283
#> [159,] 5.371056e-01 4.213894e-01 2.425601e+00 4.799426e-01  7.132684 10.561583
#> [160,] 1.594632e-02 3.253570e-02 2.478013e+00 1.353322e-01  7.346014 10.539747
#> [161,] 2.729279e-01 6.821521e-02 2.553440e+00 4.236406e-01  6.983516 10.830630
#> [162,] 2.849890e+00 3.545750e-01 2.724413e+00 8.158457e-01  7.241353 13.547705
#> [163,] 1.149840e+00 2.808120e+00 3.226788e+00 4.307843e-01  6.569790 12.913234
#> [164,] 2.036021e+00 9.795041e-01 3.547105e+00 1.590058e+00  8.520233 12.954807
#> [165,] 3.748214e-01 5.863020e-03 3.579704e+00 1.084457e-02  8.179004 11.866511
#> [166,] 3.501981e-01 2.194561e-01 3.742662e+00 4.410617e-01  7.889040 11.880547
#> [167,] 2.425260e-02 7.180144e-01 4.040639e+00 3.718792e-04  7.623164 11.894095
#> [168,] 1.970045e+00 5.591439e-01 4.315523e+00 8.020425e-01  7.391473 13.044273
#> [169,] 9.865459e-01 2.277472e-02 4.273131e+00 1.923605e+00  9.324873 13.058196
#> [170,] 3.214200e-02 6.194456e-01 4.007683e+00 9.667466e-02  6.133554 11.179744
#> [171,] 1.658839e-02 8.136154e-01 3.712142e+00 1.401093e-02  6.272350 11.374873
#> [172,] 1.085397e-02 9.695658e-01 3.401177e+00 5.503947e-01  6.472616 11.087329
#> [173,] 7.799564e-01 1.105843e+00 3.083019e+00 2.022494e-01  6.791454 11.187528
#> [174,] 3.633904e+00 2.617169e+00 2.617859e+00 2.900133e+00  6.185375 14.405235
#> [175,] 2.500977e+00 8.065553e+00 1.885136e+00 1.034066e-01  4.873890 14.405752
#> [176,] 6.084292e+00 3.336737e+00 1.478536e+00 4.411644e-01  6.260983 16.966459
#> [177,] 9.440359e-02 6.017638e-02 1.433075e+00 7.195908e-03 12.726444 15.042270
#> [178,] 3.674964e-03 1.957533e-01 1.347974e+00 1.731691e-03 12.642851 15.410637
#> [179,] 7.579226e-02 1.629375e-01 1.273174e+00 1.342906e+00 13.069496 16.121853
#> [180,] 6.972973e-01 5.156864e-02 1.234940e+00 1.646018e+00 13.733252 16.150283
#> [181,] 5.024782e-02 9.540026e-02 1.303572e+00 3.824884e-01 11.310056 14.579611
#> [182,] 1.236164e+00 2.051057e-01 1.405049e+00 6.930613e-02 11.891865 17.313065
#> [183,] 1.196290e-01 1.362169e+00 1.671859e+00 2.005225e-02 12.386121 17.049930
#> [184,] 1.039577e+00 1.931277e+00 2.020538e+00 4.155006e-05 11.119607 18.864689
#> [185,] 6.642723e-01 4.183898e+00 2.591457e+00 2.664975e-01 10.430619 20.288080
#> [186,] 1.383246e-05 6.605934e+00 3.411433e+00 2.527634e+00  9.280361 20.365188
#> [187,] 1.860057e+00 6.623384e+00 4.354536e+00 1.339322e+00  9.981976 20.080891
#> [188,] 4.015976e-02 2.880225e+00 5.060617e+00 3.820761e+00 11.717484 25.590398
#> [189,] 4.318696e-02 2.461394e+00 5.773342e+00 1.867488e-01  9.620515 18.396720
#> [190,] 5.683340e-01 2.061506e+00 6.481865e+00 2.594414e-02 10.146892 18.393660
#> [191,] 2.918038e+00 9.039166e-01 7.005694e+00 3.973773e-01 10.587954 20.146842
#> [192,] 5.326137e-01 2.251592e-02 7.012395e+00 4.987570e-01 14.016337 20.722437
#> [193,] 3.421457e-01 3.854992e-01 6.794608e+00 4.822679e-02 15.489187 20.757548
#> [194,] 3.350167e+00 9.973700e-01 6.402904e+00 9.730072e-01 14.683698 23.065440
#> [195,] 1.444219e-01 4.756801e+00 5.489906e+00 6.047814e-01 13.702687 22.440347
#> [196,] 3.221263e-01 3.750710e+00 4.740331e+00 3.485318e-02 13.801055 21.732488
#> [197,] 4.646952e-01 5.311527e+00 3.900170e+00 1.644045e+00 12.293261 22.253847
#> [198,] 2.234817e+00 7.547077e+00 2.985529e+00 8.923718e-02 11.545992 22.348055
#> [199,] 1.435221e-01 3.174091e+00 2.468666e+00 5.088222e+00 12.127352 19.093105
#> [200,] 4.875627e+00 2.364917e+00 2.063740e+00 3.848181e+00 11.727229 17.796621
#> [201,] 2.323920e+00 8.831363e+00 1.343017e+00 6.955524e-03  5.030407 17.320929
#> [202,] 1.207592e-01 3.942530e+00 9.499119e-01 5.475044e-01  9.474016 15.404858
#> [203,] 6.625727e-01 3.106191e+00 6.553665e-01 8.605923e-02  9.243332 13.513617
#> [204,] 1.534261e-01 1.525102e+00 4.830243e-01 7.106567e-01  7.109264  9.644766
#> [205,] 2.132538e-03 9.628978e-01 3.651753e-01 9.143232e-01  5.929507  6.991047
#> [206,] 1.521879e+00 1.023679e+00 2.579772e-01 1.210936e+00  5.926715  6.975508
#> [207,] 5.515105e-02 4.300490e-02 2.448226e-01 1.367690e-02  4.276269  4.810700
#> [208,] 3.498261e-02 2.939182e-03 2.483468e-01 8.668974e-02  4.509345  4.854373
#> [209,] 3.624670e-02 3.107923e-02 2.390048e-01 5.971064e-01  4.586372  4.845787
#> [210,] 3.374444e-02 2.715006e-03 2.438407e-01 5.699918e-01  4.488865  4.853229
#> [211,] 8.807459e-02 2.964651e-02 2.359353e-01 3.478398e-02  4.293946  4.663626
#> [212,] 7.788392e-01 4.667103e-04 2.516595e-01 2.439947e+00  4.221517  8.016688
#> [213,] 3.833868e+00 3.108980e-01 2.016355e-01 1.426829e+00  7.250247 10.040245
#> [214,] 2.135995e-02 3.440467e+00 5.373155e-02 8.974524e-02  4.764918  8.706381
#> [215,] 1.544829e+00 3.170038e+00 9.949687e-05 1.108734e-02  4.765058  8.498886
#> [216,] 1.601076e+00 9.734260e-01 1.508679e-02 1.849290e-01  6.261823  7.826569
#> [217,] 1.264994e-02 1.721536e-02 2.346139e-02 6.312740e-02  6.890449  6.961312
#> [218,] 6.382504e-02 5.277506e-03 3.258041e-02 2.515766e-02  6.935944  6.935944
#> [219,] 2.334241e-01 4.446278e-02 5.641378e-02 2.658603e-01  6.924738  6.924738
#> [220,] 7.082605e-02 2.306389e-01 1.297164e-01 3.796240e-03  6.808665  6.808665
#> [221,] 3.124901e-02 1.147581e-01 2.387533e-01 1.770599e-01  6.626293  6.626293
#> [222,] 1.892452e-01 5.995259e-02 4.075833e-01 9.181376e-02  6.555733  6.555733
#> [223,] 2.717690e-05 2.549584e-01 8.673364e-01 1.693540e+00  6.505575  6.505575
#> [224,] 3.617164e+00 2.708363e-01 1.918500e+00 3.103716e-01  6.488153  6.488153
#> [225,] 1.208885e+00 4.322040e-01 2.556520e+00 6.177012e-01  3.350350  3.350350
#> [226,] 2.676083e-03 2.134001e+00 2.187146e+00 5.283586e-01  2.741155  2.741155
#> [227,] 1.287196e+00 2.586485e+00 1.017524e+00 7.901872e-03  2.592685  2.592685
#> [228,] 1.017524e+00 1.017524e+00          NaN 1.017524e+00  1.017524  1.017524
#> 
#> 
#> $bsm$final
#> $bsm$final$level
#> [1] 0.4685294
#> 
#> $bsm$final$slope
#> [1] 0.0004399603
#> 
#> $bsm$final$seasonal
#> [1] 0.1859279
#> 
#> $bsm$final$noise
#> [1] 1
#> 
#> $bsm$final$tau
#>                [,1]         [,2]         [,3]         [,4]      [,5]      [,6]
#>   [1,] 4.437989e+00          NaN          NaN          NaN  0.000000  4.437989
#>   [2,] 4.746477e-01 4.437989e+00 4.4379887827          NaN  4.437989  4.637854
#>   [3,] 6.330292e-02 1.121927e+00 2.7253326642          NaN  4.637854  4.645872
#>   [4,] 1.658846e-02 5.904622e-01 1.9543982243          NaN  4.645872  4.649543
#>   [5,] 1.898129e-02 3.949707e-01 1.5154506469          NaN  4.649543  4.660143
#>   [6,] 1.926021e-01 2.675793e-01 1.2330586078          NaN  4.660143  4.837510
#>   [7,] 5.673939e-01 5.842160e-02 0.9369323072          NaN  4.837510  5.263855
#>   [8,] 4.150549e-01 3.906371e-02 0.6061775906          NaN  5.263855  5.536611
#>   [9,] 3.840397e-02 2.900095e-02 0.5004421164          NaN  5.536611  5.573762
#>  [10,] 4.341654e-01 3.431408e-03 0.4034630888          NaN  5.573762  6.081190
#>  [11,] 7.445668e-03 9.294210e-02 0.2734317705          NaN  6.081190  6.136392
#>  [12,] 6.498031e-01 1.212592e-01 0.1808061320 4.437989e+00  6.136392  6.561529
#>  [13,] 1.598805e+00 7.960633e-03 0.1600048701 3.488604e+00  6.561288  6.572989
#>  [14,] 1.969884e-01 6.064470e-01 0.0765505592 8.432285e-02  2.677765  6.573026
#>  [15,] 6.156382e-02 2.118204e-01 0.0426602450 6.947732e-03  2.331761  3.577145
#>  [16,] 2.809479e-01 3.822776e-01 0.0152351189 4.660907e-05  2.472155  3.832862
#>  [17,] 1.628301e-03 9.221735e-01 0.0000225541 4.268424e-02  2.871523  3.769427
#>  [18,] 2.886250e-01 8.659160e-01 0.0108747701 4.714738e-02  2.870077  3.592754
#>  [19,] 9.704075e-02 3.329965e-01 0.0282356977 9.224944e-01  2.956112  3.451879
#>  [20,] 2.521904e-01 1.390741e-01 0.0425973355 3.335811e-01  2.608723  4.227307
#>  [21,] 1.406043e-01 1.993598e-03 0.0429813095 1.013530e-01  3.968896  4.080878
#>  [22,] 4.533543e-05 4.010317e-02 0.0330797243 1.550567e-01  3.983568  4.218321
#>  [23,] 5.264814e-02 3.833653e-02 0.0249451007 3.766124e-01  3.892832  4.142688
#>  [24,] 1.272711e-02 2.113929e-03 0.0226777082 1.032996e-01  4.079052  4.490117
#>  [25,] 2.444051e+00 7.645745e-04 0.0228555299 7.939112e-02  3.896100 10.182818
#>  [26,] 1.436037e-02 9.840710e-01 0.0022938850 1.222496e-01  9.362625 10.204669
#>  [27,] 1.480989e-02 8.317776e-01 0.0020085965 3.184265e-02  9.270236 10.258059
#>  [28,] 2.012571e-01 6.912249e-01 0.0162623414 2.719043e-01  9.054153 10.075189
#>  [29,] 1.682774e-02 1.257147e+00 0.0564046532 6.543124e-02  8.258618  9.943838
#>  [30,] 5.103108e-02 1.449551e+00 0.1251931142 1.810853e-01  8.191328  9.930001
#>  [31,] 1.280545e-01 1.116155e+00 0.2061762453 1.286719e+00  8.268346  9.783524
#>  [32,] 3.916252e-01 1.656741e+00 0.3310881620 4.813307e-01  8.091054 10.102485
#>  [33,] 8.479698e-02 7.773276e-01 0.4306207543 4.519489e-01  7.046978  8.237434
#>  [34,] 3.903568e-03 4.802444e-01 0.5157669596 3.440877e-02  7.027026  8.282192
#>  [35,] 1.877257e-01 4.254613e-01 0.6018052394 5.188321e-01  7.269738  8.154136
#>  [36,] 1.314652e-01 8.678868e-01 0.7386164026 1.537318e+00  7.181453  8.174379
#>  [37,] 2.112062e+00 4.861274e-01 0.8471336735 3.419373e+00  6.917936  7.885995
#>  [38,] 5.519018e-03 5.769804e-02 0.7977047780 1.234501e-01  1.685309  2.949773
#>  [39,] 1.844904e-01 8.290939e-02 0.7430203796 1.243804e-01  1.692501  2.728502
#>  [40,] 1.942034e-02 1.243414e-04 0.7346716839 5.037632e-02  1.230656  1.993901
#>  [41,] 1.061379e-02 6.171953e-03 0.7411195153 5.686902e-03  1.181853  2.003320
#>  [42,] 2.031379e-01 2.097068e-02 0.7585756592 1.395719e-02  1.222338  2.340294
#>  [43,] 5.013201e-01 2.106011e-02 0.7301352045 4.760353e-02  1.671787  2.924803
#>  [44,] 3.544488e-03 3.603122e-01 0.6344032023 1.460253e-01  2.074650  3.061054
#>  [45,] 1.702706e+00 3.155992e-01 0.5513856069 2.108959e-01  2.106092  6.881951
#>  [46,] 1.911105e+00 1.959778e+00 0.3751477880 4.188933e-01  4.885828  7.836310
#>  [47,] 3.745155e-02 2.617514e-01 0.3183135453 4.388563e-04  6.185055  6.947807
#>  [48,] 4.025523e-01 4.041469e-01 0.2549674133 5.017881e-02  6.108084  8.074954
#>  [49,] 8.336748e-01 5.211510e-02 0.2333067477 3.189785e-01  7.551508 10.677185
#>  [50,] 3.155482e-02 1.279548e-01 0.2645315308 1.389711e-02 10.016857 10.803718
#>  [51,] 2.864385e-02 2.223913e-01 0.3089920926 6.793294e-03  9.866616 10.734699
#>  [52,] 1.004097e-02 3.364323e-01 0.3685865176 3.633081e-02  9.659444 10.639665
#>  [53,] 2.142937e-01 2.658841e-01 0.4256856323 3.494546e-01  9.647804 10.658970
#>  [54,] 3.505125e-02 4.784585e-02 0.4502338459 1.477071e-02  9.827583 10.658876
#>  [55,] 1.447352e-01 1.147202e-01 0.4904650114 8.706563e-01  9.617262 10.450590
#>  [56,] 3.012675e+00 3.392632e-01 0.5647794495 2.732175e+00  8.793320 12.809611
#>  [57,] 4.127153e+00 2.807384e-01 0.4941826826 4.176660e+00 11.770094 13.108598
#>  [58,] 1.321482e+00 5.956809e-01 0.5946083051 6.733881e-01  9.573767 10.748250
#>  [59,] 1.839319e-01 1.246110e-03 0.5979922026 5.628534e-01  8.904730  9.479255
#>  [60,] 4.851006e-01 9.609067e-02 0.6401150763 2.337530e-01  9.121875 12.535244
#>  [61,] 8.171778e+00 5.715081e-01 0.7510477082 1.561862e+00 11.061231 16.289015
#>  [62,] 2.972844e-01 1.154703e+00 0.5928415959 1.211541e-02 11.580646 13.904141
#>  [63,] 8.073918e-02 5.261393e-01 0.4964926263 2.840614e-02 12.330090 13.748935
#>  [64,] 7.335082e-04 2.952498e-01 0.4298540062 2.167435e-02 12.581745 13.676291
#>  [65,] 3.551618e-04 3.143263e-01 0.3662958998 2.069625e-04 12.542404 13.527711
#>  [66,] 7.452422e-02 3.008856e-01 0.3091378271 3.414036e-03 12.497075 13.194689
#>  [67,] 7.689485e-02 1.397025e-01 0.2728974216 1.077180e+00 12.387040 12.844849
#>  [68,] 1.970686e+00 3.852676e-02 0.2546307263 3.444197e+00 12.308877 12.829005
#>  [69,] 5.943838e-02 4.928121e-01 0.3209591364 8.448597e-01  9.537438 10.462289
#>  [70,] 1.448868e-02 2.980732e-01 0.3776990634 4.218976e-01  9.204615  9.953904
#>  [71,] 1.037295e+00 2.198775e-01 0.4300142521 9.989885e-01  9.500285 10.681723
#>  [72,] 1.136424e+00 3.337327e-02 0.4086003630 7.996115e+00 10.254422 10.630878
#>  [73,] 1.605940e+00 7.475576e-01 0.3162523683 3.688700e+00  6.339845  7.573531
#>  [74,] 1.473292e-01 2.910023e-03 0.3106041586 1.771800e-02  2.510596  2.719345
#>  [75,] 2.232206e-03 8.967196e-02 0.2815185703 2.101478e-01  2.057097  2.599816
#>  [76,] 7.634448e-02 7.247860e-02 0.2566003347 1.090855e-02  2.095575  2.601783
#>  [77,] 4.702358e-02 8.554588e-03 0.2481843958 6.304922e-02  2.122352  2.473852
#>  [78,] 1.150142e-01 2.134001e-03 0.2520794398 3.899030e-03  1.989018  2.599676
#>  [79,] 7.608006e-01 6.920730e-02 0.2757634628 1.289417e-04  2.211467  4.627488
#>  [80,] 4.455013e-01 6.737764e-01 0.3569233701 1.032192e-01  3.575853  4.866083
#>  [81,] 1.162695e+00 1.552351e-01 0.3994522464 6.467368e-01  4.195346  5.315218
#>  [82,] 8.823735e-02 8.732122e-02 0.3669478851 3.495308e-02  4.619980  5.094700
#>  [83,] 9.292430e-01 1.114228e-02 0.3555551583 9.124093e-01  4.378206  5.089359
#>  [84,] 1.496562e-01 2.609147e-01 0.4111196448 4.284018e-01  3.742243  5.639371
#>  [85,] 2.967672e-01 5.747535e-01 0.5011049329 1.465193e-01  4.306825  6.211201
#>  [86,] 2.637147e-02 1.679435e-01 0.5533286260 7.203763e-02  5.592846  6.398589
#>  [87,] 2.680654e-01 2.638071e-01 0.6224905312 6.006140e-02  5.648491  6.405994
#>  [88,] 3.269420e-01 3.334348e-02 0.6478950716 2.554273e-02  5.543698  6.357135
#>  [89,] 4.712191e-01 3.347005e-02 0.6219678408 1.332212e-01  5.810418  6.986863
#>  [90,] 1.446028e-02 6.548393e-02 0.6578192232 3.124488e-01  6.337478  7.103591
#>  [91,] 2.149257e+00 3.204800e-02 0.6834348097 2.024116e+00  6.246265  7.498011
#>  [92,] 6.308483e+00 5.748172e-01 0.5772871297 3.765465e-03  5.266175 14.603488
#>  [93,] 3.383867e+00 7.181516e-01 0.6960984858 2.273190e-01 11.674320 17.668173
#>  [94,] 2.030092e-01 4.093861e+00 1.0249516770 1.885421e-01 11.751272 16.555032
#>  [95,] 5.503996e-01 3.011204e+00 1.3574021455 1.744524e-01 12.196036 16.550058
#>  [96,] 4.782168e+00 1.590182e+00 1.6281073693 2.317185e+00 12.793675 18.741201
#>  [97,] 2.942585e+00 1.872487e-02 1.5971726693 1.084559e+00 15.679626 19.474901
#>  [98,] 5.798537e-01 9.209416e-01 1.8175088665 1.332873e-01 15.206823 18.353764
#>  [99,] 9.489730e-03 2.236412e-01 1.9311469723 8.726425e-02 16.150388 18.251026
#> [100,] 4.786736e-02 2.864064e-01 2.0639559845 1.024352e+00 16.138548 18.999294
#> [101,] 1.608698e+00 1.562771e-01 2.1648183296 1.492802e-01 17.352896 19.452153
#> [102,] 9.671245e-01 1.725467e-01 2.0582905672 4.823137e-01 17.017614 20.475740
#> [103,] 1.325004e-01 1.089866e+00 1.8028811355 5.208022e+00 18.125804 22.479025
#> [104,] 4.110640e-01 1.629762e+00 1.5136359446 4.725095e-01 19.231562 29.688022
#> [105,] 2.247358e+00 7.513686e-01 1.3316302037 2.743026e+00 23.343415 27.037313
#> [106,] 1.581142e-02 3.330602e+00 0.9867040480 2.652863e-02 20.055740 26.196368
#> [107,] 1.061803e-01 3.043678e+00 0.7052426286 8.614202e-01 19.581791 25.357440
#> [108,] 2.445061e+00 3.813732e+00 0.4461383221 4.461736e+00 19.989680 24.439894
#> [109,] 1.035050e+01 9.090523e-01 0.3410916896 1.601582e+00 17.766200 47.591537
#> [110,] 6.621447e-01 1.216296e+00 0.4638238347 1.891283e-01 45.193667 48.757979
#> [111,] 1.870059e+00 3.396123e-01 0.5362688212 3.479990e-01 46.724596 49.793906
#> [112,] 5.059707e-01 8.484272e-02 0.4993787154 1.193449e-03 48.956732 49.798693
#> [113,] 3.306977e-01 2.668656e-02 0.5198879044 1.463965e-02 48.117277 49.027746
#> [114,] 6.455887e-02 2.818668e-01 0.5894199063 1.382102e-02 47.112048 48.548267
#> [115,] 7.527044e-01 4.806772e-01 0.6867965003 1.035747e+01 47.106693 48.155041
#> [116,] 5.886465e+00 1.557069e+00 0.8808119115 7.186841e+00 45.851782 47.478622
#> [117,] 1.084174e+00 9.173685e-02 0.8315214084 2.039612e-02 35.705132 38.022463
#> [118,] 9.926157e-03 1.314975e-01 0.8907480050 6.642995e-02 36.779615 38.037025
#> [119,] 6.619873e-03 1.817362e-01 0.9629872536 6.918407e-01 36.559471 38.152046
#> [120,] 1.476449e-01 1.401029e-01 1.0287536144 1.333999e+01 37.381424 38.164090
#> [121,] 8.483434e+00 1.656664e-02 1.0519151750 2.519987e+01 37.709281 39.348929
#> [122,] 6.291525e-01 3.003017e+00 0.7624595306 6.544660e-03 11.027422 15.980873
#> [123,] 1.287187e-02 1.502959e+00 0.5857634558 1.812501e+00  8.493725 13.265350
#> [124,] 3.414422e-01 1.330430e+00 0.4407504586 8.530254e-03  7.799087 10.371660
#> [125,] 3.159689e-01 2.331530e+00 0.2804306554 1.763609e-01  8.157682 11.738679
#> [126,] 1.295955e+00 3.557866e+00 0.1322008927 2.059804e-01  8.388516 11.862969
#> [127,] 2.063258e-01 1.342366e+00 0.0684615731 2.264245e-02  9.149407 10.833399
#> [128,] 8.019495e+00 7.539072e-01 0.0343176773 3.011138e-01  9.537799 26.557741
#> [129,] 2.085711e-01 8.870078e-01 0.0719102219 1.040044e+00 25.785046 26.767549
#> [130,] 3.108880e-02 4.223737e-01 0.1058788276 1.817456e-01 25.578462 26.346094
#> [131,] 4.181308e-02 2.885900e-01 0.1389129847 3.021313e-01 25.157405 27.139125
#> [132,] 4.924296e-01 1.652464e-01 0.1669005453 1.683404e+00 25.494505 27.134990
#> [133,] 5.951279e-01 7.311424e-01 0.2341240330 2.459721e+00 25.825906 27.376304
#> [134,] 7.960191e-01 1.310106e-01 0.2660396279 5.098027e-01 24.768437 27.053722
#> [135,] 5.177363e+00 8.691493e-01 0.3575496259 4.849399e-01 24.561525 30.187175
#> [136,] 6.549553e-02 5.696894e+00 0.6532721179 5.135759e-03 20.167058 27.354510
#> [137,] 5.292041e-03 6.504816e+00 1.0671423647 1.354661e+00 19.761276 27.368817
#> [138,] 7.352242e-01 6.744426e+00 1.5925725951 3.362931e-02 19.777193 27.276668
#> [139,] 4.189720e-02 4.198006e+00 2.0813964773 5.593239e+00 19.775130 24.964356
#> [140,] 1.066262e+01 3.679070e+00 2.5984874789 6.321331e+00 19.744625 24.381923
#> [141,] 4.571163e-01 2.870527e-02 2.5515151764 5.756016e-01  5.730011  8.526149
#> [142,] 2.200562e-01 3.620131e-01 2.3858535602 2.047984e-01  5.799447  8.510563
#> [143,] 1.972471e+00 9.106986e-02 2.3053499813 1.942433e+00  5.441342  8.237819
#> [144,] 9.756876e-01 3.553920e-01 2.4688313263 5.573295e-03  4.054799 10.772410
#> [145,] 1.553617e+00 1.507278e+00 2.8221303369 3.547706e-04  6.692589 11.279036
#> [146,] 5.816759e-01 1.855803e-01 2.9527482125 3.298073e-01  8.045211 11.019263
#> [147,] 5.200818e-01 3.232745e-03 2.9371362487 1.403147e+00  8.042718 11.023449
#> [148,] 4.764278e-02 2.683184e-01 2.7841685754 5.869597e-02  6.564044 10.336528
#> [149,] 2.047080e-01 4.324222e-01 2.5954474396 8.851480e-02  6.691567  9.894582
#> [150,] 2.930206e-01 1.356333e-01 2.4933834375 1.802850e-01  6.242582  9.749234
#> [151,] 3.532142e-03 5.104697e-01 2.2997113097 2.359089e-01  7.143621  9.594057
#> [152,] 1.359978e+00 5.662737e-01 2.1040807247 1.052543e-02  7.050781 11.015068
#> [153,] 4.053813e-03 4.491976e-05 2.1040454495 9.483749e-02  8.290183 10.797094
#> [154,] 7.512523e-02 2.249148e-03 2.0936840053 1.752524e-01  8.393108 10.448872
#> [155,] 3.883991e-08 4.961445e-02 2.0389166136 1.011270e+00  8.574098 10.646823
#> [156,] 1.134758e+00 4.956391e-02 1.9849744868 4.646576e+00  8.522090 10.809572
#> [157,] 4.792967e+00 8.173698e-01 1.7677199811 1.864010e-01  5.286804 13.190946
#> [158,] 5.431119e-02 2.464480e-01 1.8886153517 3.757086e-01  7.704410 10.736060
#> [159,] 5.567752e-01 4.167601e-01 2.0514772987 5.128712e-01  7.091131 10.105597
#> [160,] 1.699744e-02 2.828807e-02 2.0970139617 1.462550e-01  7.304056 10.077857
#> [161,] 2.743310e-01 6.331642e-02 2.1650123339 3.944730e-01  6.932571 10.307383
#> [162,] 2.696721e+00 3.443620e-01 2.3242451762 7.628680e-01  7.172572 12.850690
#> [163,] 1.097917e+00 2.682539e+00 2.7930164674 4.301091e-01  6.459107 12.209031
#> [164,] 1.841029e+00 9.357573e-01 3.0926404546 1.557684e+00  8.305053 12.190275
#> [165,] 3.629911e-01 9.782875e-03 3.1289902888 1.253780e-02  7.842543 11.088896
#> [166,] 3.392061e-01 2.348553e-01 3.2893864061 4.544127e-01  7.542686 11.095818
#> [167,] 2.501832e-02 7.354157e-01 3.5791621342 1.759835e-04  7.281995 11.130557
#> [168,] 2.057244e+00 5.721169e-01 3.8467272761 7.667960e-01  7.037450 12.306074
#> [169,] 9.059116e-01 2.626427e-02 3.7985865732 1.932905e+00  8.945204 12.285702
#> [170,] 2.589146e-02 5.955223e-01 3.5430413659 8.945590e-02  5.878013 10.468459
#> [171,] 2.160620e-02 7.654248e-01 3.2629121101 1.005260e-02  6.003003 10.658733
#> [172,] 7.060552e-03 9.393326e-01 2.9650126146 5.607558e-01  6.222611 10.388236
#> [173,] 7.358351e-01 1.046920e+00 2.6652999815 1.710688e-01  6.518106 10.449162
#> [174,] 3.535001e+00 2.474742e+00 2.2312207835 2.663475e+00  5.938991 13.793318
#> [175,] 2.375734e+00 7.720752e+00 1.5529247868 1.354227e-01  4.894893 13.774604
#> [176,] 6.048831e+00 3.207541e+00 1.1814510782 3.327445e-01  6.149843 16.643411
#> [177,] 8.259971e-02 4.586908e-02 1.1440944086 8.700393e-03 12.901752 14.838722
#> [178,] 5.945831e-03 1.588520e-01 1.0726725788 2.457365e-04 12.834044 15.195172
#> [179,] 5.875024e-02 1.219312e-01 1.0125433337 1.381380e+00 13.245706 15.835332
#> [180,] 6.595080e-01 3.753101e-02 0.9819497505 1.755841e+00 13.833960 15.859851
#> [181,] 4.000148e-02 1.072596e-01 1.0467408282 4.272762e-01 11.455867 14.370885
#> [182,] 1.209090e+00 2.079465e-01 1.1391294043 8.186972e-02 12.026674 16.985058
#> [183,] 1.310750e-01 1.351081e+00 1.3840285468 2.525122e-02 12.492477 16.722829
#> [184,] 1.068428e+00 1.946902e+00 1.7101749814 7.920049e-04 11.210948 18.549244
#> [185,] 6.153380e-01 4.243870e+00 2.2533314945 2.900951e-01 10.549800 19.942326
#> [186,] 5.845525e-05 6.579838e+00 3.0364045438 2.417222e+00  9.474907 20.014828
#> [187,] 1.783939e+00 6.610248e+00 3.9460825834 1.495981e+00 10.131230 19.797033
#> [188,] 3.930297e-02 2.933565e+00 4.6354605918 3.926501e+00 11.863803 25.264874
#> [189,] 3.667681e-02 2.515227e+00 5.3341656445 1.872966e-01  9.805409 18.128781
#> [190,] 5.714035e-01 2.141569e+00 6.0351213627 1.911573e-02 10.260600 18.130709
#> [191,] 2.909977e+00 9.547135e-01 6.5542809169 4.156297e-01 10.670323 19.855162
#> [192,] 5.160793e-01 1.493933e-02 6.5628406001 5.699958e-01 14.066791 20.434789
#> [193,] 3.530532e-01 3.431087e-01 6.3529248783 3.392865e-02 15.574028 20.457292
#> [194,] 3.256425e+00 9.400527e-01 5.9686072720 9.468634e-01 14.748734 22.712955
#> [195,] 1.585013e-01 4.559883e+00 5.0801486872 6.153310e-01 13.812314 22.089080
#> [196,] 3.175536e-01 3.531805e+00 4.3597204723 4.160473e-02 13.917511 21.346203
#> [197,] 4.318195e-01 5.038778e+00 3.5542814757 1.592885e+00 12.467100 21.864161
#> [198,] 2.240888e+00 7.136571e+00 2.6859146511 9.219022e-02 11.781877 21.965524
#> [199,] 1.398501e-01 2.904810e+00 2.2048076591 5.209583e+00 12.274627 18.655856
#> [200,] 5.077759e+00 2.142504e+00 1.8310975318 3.901673e+00 11.880558 17.569538
#> [201,] 2.269323e+00 8.570849e+00 1.1522789516 8.628902e-03  4.962469 16.773850
#> [202,] 1.142807e-01 3.814680e+00 0.7881345505 5.572690e-01  9.255665 14.846842
#> [203,] 6.840028e-01 3.015033e+00 0.5200496712 7.563695e-02  9.034663 13.012848
#> [204,] 1.722497e-01 1.441150e+00 0.3682532335 7.580751e-01  6.929186  9.217832
#> [205,] 2.303293e-03 8.680789e-01 0.2687848164 8.894880e-01  5.713734  6.610609
#> [206,] 1.445505e+00 9.281093e-01 0.1800300161 1.163080e+00  5.708308  6.609895
#> [207,] 6.002103e-02 3.192086e-02 0.1699835641 1.236822e-02  4.128361  4.587970
#> [208,] 3.740162e-02 3.503488e-04 0.1745911029 8.477316e-02  4.372803  4.630188
#> [209,] 3.258301e-02 2.104447e-02 0.1679189035 5.537370e-01  4.446332  4.626836
#> [210,] 2.415022e-02 7.393249e-04 0.1726070842 5.510238e-01  4.358587  4.658810
#> [211,] 9.160131e-02 1.661043e-02 0.1681970913 2.703631e-02  4.201448  4.484490
#> [212,] 8.335374e-01 4.775529e-03 0.1843634912 2.329552e+00  4.122036  7.829061
#> [213,] 3.711173e+00 2.818105e-01 0.1421854122 1.394843e+00  7.188804  9.953589
#> [214,] 2.440472e-02 3.275152e+00 0.0260828030 8.478282e-02  4.717387  8.586537
#> [215,] 1.555245e+00 2.988665e+00 0.0033308027 1.180837e-02  4.724267  8.377537
#> [216,] 1.623039e+00 8.617362e-01 0.0350321336 1.938907e-01  6.194924  7.694509
#> [217,] 8.246824e-03 3.724545e-03 0.0455681602 4.839203e-02  6.786901  6.820595
#> [218,] 5.247589e-02 1.527711e-04 0.0564104510 3.571563e-02  6.806319  6.806319
#> [219,] 2.539319e-01 1.921075e-02 0.0832494429 2.708127e-01  6.795568  6.795568
#> [220,] 6.392285e-02 1.776908e-01 0.1650257109 3.206891e-03  6.651935  6.651935
#> [221,] 2.912599e-02 8.096624e-02 0.2802387991 1.878514e-01  6.494215  6.494215
#> [222,] 2.115946e-01 3.676277e-02 0.4521567808 1.012297e-01  6.435099  6.435099
#> [223,] 6.886484e-06 2.182019e-01 0.9222561635 1.688014e+00  6.365863  6.365863
#> [224,] 3.602697e+00 2.324572e-01 1.9780917531 3.529256e-01  6.352240  6.352240
#> [225,] 1.082932e+00 4.933090e-01 2.5754464705 5.661917e-01  3.267854  3.267854
#> [226,] 3.959418e-03 2.164581e+00 2.1893523299 5.135684e-01  2.759488  2.759488
#> [227,] 1.282395e+00 2.591207e+00 1.0145216730 7.860630e-03  2.597234  2.597234
#> [228,] 1.014522e+00 1.014522e+00          NaN 1.014522e+00  1.014522  1.014522
#> 
#> 
#> 
#> $likelihood
#> $likelihood$initial
#> $likelihood$initial$ll
#> [1] -1233.697
#> 
#> $likelihood$initial$ssq
#> [1] 34176.74
#> 
#> $likelihood$initial$nobs
#> [1] 228
#> 
#> $likelihood$initial$neffectiveobs
#> [1] -1
#> 
#> $likelihood$initial$nparams
#> [1] 4
#> 
#> $likelihood$initial$df
#> [1] 211
#> 
#> $likelihood$initial$aic
#> [1] 2475.393
#> 
#> $likelihood$initial$aicc
#> [1] 2475.583
#> 
#> $likelihood$initial$bic
#> [1] 2488.876
#> 
#> $likelihood$initial$bic2
#> [1] NaN
#> 
#> $likelihood$initial$bicc
#> [1] NaN
#> 
#> $likelihood$initial$hannanquinn
#> [1] 2480.841
#> 
#> 
#> $likelihood$final
#> $likelihood$final$ll
#> [1] -1233.634
#> 
#> $likelihood$final$ssq
#> [1] 58486.05
#> 
#> $likelihood$final$nobs
#> [1] 228
#> 
#> $likelihood$final$neffectiveobs
#> [1] -1
#> 
#> $likelihood$final$nparams
#> [1] 4
#> 
#> $likelihood$final$df
#> [1] 211
#> 
#> $likelihood$final$aic
#> [1] 2475.267
#> 
#> $likelihood$final$aicc
#> [1] 2475.458
#> 
#> $likelihood$final$bic
#> [1] 2488.75
#> 
#> $likelihood$final$bic2
#> [1] NaN
#> 
#> $likelihood$final$bicc
#> [1] NaN
#> 
#> $likelihood$final$hannanquinn
#> [1] 2480.715
#> 
#> 
#> 
#> attr(,"class")
#> [1] "JD3_STS_OUTLIERS"